refactoring lokidb testing files. next step: add list blob files
