+++ tests/blob/versioning.lokidb.test.ts
@@ -1,1947 +1,1390 @@
 import assert = require("assert");
 import { v4 as uuid } from "uuid";
-import * as fs from "fs";
 import LokiBlobMetadataStore from "../../src/blob/persistence/LokiBlobMetadataStore";
 import {
-  BlobModel,
-  ContainerModel
-} from "../../src/blob/persistence/IBlobMetadataStore";
+  buildAppendBlob,
+  buildBlockBlob,
+  buildContainer,
+  buildPageBlob,
+  createContext
+} from "../testutils";
 import * as Models from "../../src/blob/generated/artifacts/models";
 import Context from "../../src/blob/generated/Context";
 import { configLogger } from "../../src/common/Logger";
 import { isNullOrWhitespace } from "../../src/blob/utils/utils";
+
 // Silence logs for tests
 configLogger(false);
 
-/**
- * Helper to create a minimal Context object.
- */
-function createContext(): Context {
-  return {
-    contextId: uuid(),
-    startTime: new Date()
-  } as any as Context; // Cast to simplify test construction
-}
-
-/**
- * Helper to build a minimal ContainerModel for tests.
- */
-function buildContainer(account: string, name: string): ContainerModel {
-  const now = new Date();
-  return {
-    accountName: account,
-    name,
-    properties: {
-      lastModified: now,
-      etag: '"test-etag"',
-      leaseStatus: Models.LeaseStatusType.Unlocked,
-      leaseState: Models.LeaseStateType.Available,
-      defaultEncryptionScope: undefined,
-      denyEncryptionScopeOverride: undefined,
-      hasImmutabilityPolicy: undefined,
-      hasLegalHold: undefined,
-      publicAccess: undefined,
-      leaseDuration: undefined
-    }
-  } as any as ContainerModel;
-}
-
-/**
- * Helper to build a minimal Block Blob BlobModel for tests.
- */
-function buildBlockBlob(
-  account: string,
-  container: string,
-  name: string,
-  content: string
-): BlobModel {
-  const now = new Date();
-  return {
-    accountName: account,
-    containerName: container,
-    name,
-    properties: {
-      creationTime: now,
-      lastModified: now,
-      etag: `\"etag-${uuid()}\"`,
-      blobType: Models.BlobType.BlockBlob,
-      contentLength: Buffer.byteLength(content),
-      serverEncrypted: false,
-      accessTier: Models.AccessTier.Hot,
-      accessTierInferred: true,
-      cacheControl: undefined,
-      contentType: undefined,
-      contentMD5: undefined,
-      contentEncoding: undefined,
-      contentLanguage: undefined,
-      contentDisposition: undefined,
-      leaseDuration: undefined,
-      leaseState: Models.LeaseStateType.Available,
-      leaseStatus: Models.LeaseStatusType.Unlocked,
-      tagCount: undefined,
-      archiveStatus: undefined,
-      accessTierChangeTime: undefined,
-      deletedTime: undefined,
-      remainingRetentionDays: undefined,
-      deleted: false,
-      rehydratePriority: undefined,
-      lastAccessedOn: undefined,
-      snapshot: undefined
-    },
-    isCommitted: true,
-    committedBlocksInOrder: [],
-    // Versioning top-level fields (duplicated when persisted in Loki)
-    snapshot: ""
-  } as any as BlobModel;
-}
-
-/**
- * Helper to build a minimal Page Blob BlobModel for tests.
- */
-function buildPageBlob(
-  account: string,
-  container: string,
-  name: string,
-  contentLength: number
-): BlobModel {
-  const now = new Date();
-  return {
-    accountName: account,
-    containerName: container,
-    name,
-    properties: {
-      creationTime: now,
-      lastModified: now,
-      etag: `\"etag-${uuid()}\"`,
-      blobType: Models.BlobType.PageBlob,
-      contentLength,
-      serverEncrypted: false,
-      accessTier: undefined,
-      accessTierInferred: undefined,
-      cacheControl: undefined,
-      contentType: undefined,
-      contentMD5: undefined,
-      contentEncoding: undefined,
-      contentLanguage: undefined,
-      contentDisposition: undefined,
-      leaseDuration: undefined,
-      leaseState: Models.LeaseStateType.Available,
-      leaseStatus: Models.LeaseStatusType.Unlocked,
-      tagCount: undefined,
-      archiveStatus: undefined,
-      accessTierChangeTime: undefined,
-      deletedTime: undefined,
-      remainingRetentionDays: undefined,
-      deleted: false,
-      rehydratePriority: undefined,
-      lastAccessedOn: undefined,
-      snapshot: undefined,
-      blobSequenceNumber: 0
-    },
-    isCommitted: true,
-    pageRangesInOrder: [],
-    snapshot: ""
-  } as any as BlobModel;
-}
-
-/**
- * Helper to build a minimal Append Blob BlobModel for tests.
- */
-function buildAppendBlob(
-  account: string,
-  container: string,
-  name: string
-): BlobModel {
-  const now = new Date();
-  return {
-    accountName: account,
-    containerName: container,
-    name,
-    properties: {
-      creationTime: now,
-      lastModified: now,
-      etag: `\"etag-${uuid()}\"`,
-      blobType: Models.BlobType.AppendBlob,
-      contentLength: 0,
-      serverEncrypted: false,
-      accessTier: undefined,
-      accessTierInferred: undefined,
-      cacheControl: undefined,
-      contentType: undefined,
-      contentMD5: undefined,
-      contentEncoding: undefined,
-      contentLanguage: undefined,
-      contentDisposition: undefined,
-      leaseDuration: undefined,
-      leaseState: Models.LeaseStateType.Available,
-      leaseStatus: Models.LeaseStatusType.Unlocked,
-      tagCount: undefined,
-      archiveStatus: undefined,
-      accessTierChangeTime: undefined,
-      deletedTime: undefined,
-      remainingRetentionDays: undefined,
-      deleted: false,
-      rehydratePriority: undefined,
-      lastAccessedOn: undefined,
-      snapshot: undefined,
-      isSealed: false
-    },
-    isCommitted: true,
-    committedBlocksInOrder: [],
-    snapshot: ""
-  } as any as BlobModel;
-}
-
 const ACCOUNT = "devstoreaccount1";
 
-describe("LokiBlobMetadataStoreVersioning", () => {
-  describe("When blob versioning disabled", () => {
-    let store: LokiBlobMetadataStore;
-    let containerName: string;
-    let ctx: Context;
-    const DB_FILE = "__test_db_blob__.json"; // standard shared test db path
-    let originalDbContent: string | undefined;
-    let originalExists = false;
+describe("LokiBlobMetadataStore - Versioning Enabled", () => {
+  let store: LokiBlobMetadataStore;
+  let containerName: string;
+  let ctx: Context;
+  const DB_FILE = "__test_db_blob__.json"; // standard shared test db path
 
-    before(() => {
-      if (fs.existsSync(DB_FILE)) {
-        originalExists = true;
-        originalDbContent = fs.readFileSync(DB_FILE, "utf8");
-      }
+  beforeEach(async () => {
+    ctx = createContext();
+    containerName = `container-${uuid()}`;
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+    await store.createContainer(ctx, buildContainer(ACCOUNT, containerName));
+  });
+
+  afterEach(async () => {
+    await store.close();
+    await store.clean();
+  });
+
+  // ================== VERSION MODE TRANSITION TESTS (ENABLED → DISABLED) ==================
+  it("should handle setBlobMetadata correctly when disabling versioning after creating versions @loki", async () => {
+    // Close the in-memory disabled store from beforeEach; we need persistence for this scenario
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned blob
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    const createdBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
+
+    // Set metadata to create versions (should create version)
+    ctx.startTime = new Date(Date.now() + 100);
+    const modifiedMetadataBaseBlob = await enabledStore.setBlobMetadata(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { versionedmeta: "value1" }
+    );
+    assert.ok(!isNullOrWhitespace(createdBaseBlob.versionId));
+    assert.ok(!isNullOrWhitespace(modifiedMetadataBaseBlob.versionId));
+    assert.notStrictEqual(
+      modifiedMetadataBaseBlob.versionId,
+      createdBaseBlob.versionId
+    );
+
+    const versionedFetched = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(!isNullOrWhitespace(versionedFetched.versionId));
+    assert.deepStrictEqual(versionedFetched.metadata, {
+      versionedmeta: "value1"
+    });
+    assert.strictEqual(
+      versionedFetched.versionId,
+      modifiedMetadataBaseBlob.versionId
+    );
+    const versionId = versionedFetched.versionId;
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Set metadata should NOT create new version (overwrite current)
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setBlobMetadata(ctx, ACCOUNT, containerName, name, undefined, {
+      disabledmeta: "value2"
     });
 
-    beforeEach(async () => {
-      ctx = createContext();
-      containerName = `container-${uuid()}`;
-      // Use in-memory for regular tests (fast); special test will override
-      store = new LokiBlobMetadataStore(DB_FILE, true, false);
-      await store.init();
-      await store.createContainer(ctx, buildContainer(ACCOUNT, containerName));
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Should be same version (no new version created)
+    assert.strictEqual(current.versionId, "");
+    assert.notStrictEqual(current.versionId, versionId);
+    assert.deepStrictEqual(current.metadata, { disabledmeta: "value2" });
+
+    const firstVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      versionId
+    );
+    assert.strictEqual(firstVersion.versionId, versionId);
+  });
+
+  it("should handle setBlobHTTPHeaders correctly when disabling versioning after creating versions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned blob
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await enabledStore.createBlob(ctx, baseBlob);
+
+    // Set HTTP headers (should NOT create version even when versioning enabled)
+    ctx.startTime = new Date(Date.now() + 100);
+    await enabledStore.setBlobHTTPHeaders(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { blobContentType: "text/plain" }
+    );
+
+    const versionedFetched = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(!isNullOrWhitespace(versionedFetched.versionId));
+    assert.strictEqual(versionedFetched.properties.contentType, "text/plain");
+    const versionId = versionedFetched.versionId;
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Set headers should continue to NOT create version and update in place
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setBlobHTTPHeaders(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { blobContentType: "application/json" }
+    );
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should be same version (headers don't create versions in either mode)
+    assert.strictEqual(current.versionId, versionId);
+    assert.strictEqual(current.properties.contentType, "application/json");
+  });
+
+  it("should handle setBlobTag/getBlobTag correctly when disabling versioning after creating versions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned blob
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    const creaatedBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
+
+    // Set tags (should NOT create version even when versioning enabled)
+    ctx.startTime = new Date(Date.now() + 100);
+    await enabledStore.setBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined,
+      { blobTagSet: [{ key: "env", value: "test" }] }
+    );
+
+    const versionedFetched = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const versionId = versionedFetched.versionId;
+    assert.ok(!isNullOrWhitespace(versionId));
+    assert.strictEqual(creaatedBaseBlob.versionId, versionId);
+    const versionedTags = await enabledStore.getBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      versionId,
+      undefined
+    );
+    assert.deepStrictEqual(versionedTags, {
+      blobTagSet: [{ key: "env", value: "test" }]
     });
-
-    afterEach(async () => {
-      if (store) {
-        await store.close();
-        await store.clean();
-      }
-    });
-
-    after(() => {
-      // Restore DB file to its original state
-      if (originalExists) {
-        fs.writeFileSync(DB_FILE, originalDbContent!);
-      } else if (fs.existsSync(DB_FILE)) {
-        try {
-          fs.unlinkSync(DB_FILE);
-        } catch {
-          /* ignore */
-        }
-      }
-    });
-
-    it("creates base blob with versionId == '' and treats it as latest when no version specified @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const contentV1 = "content-v1";
-      const blobV1 = buildBlockBlob(ACCOUNT, containerName, name, contentV1);
-      await store.createBlob(ctx, blobV1);
-
-      // Fetch without versionId (downloadBlob) => latest
-      const fetched = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(
-        fetched.versionId,
-        "",
-        "Base version should have empty versionId"
-      );
-    });
-
-    it("overwrites base blob and keeps latest (no version) when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blobV1 = buildBlockBlob(ACCOUNT, containerName, name, "bob");
-      const createdV1 = await store.createBlob(ctx, blobV1);
-
-      const blobV2 = buildBlockBlob(ACCOUNT, containerName, name, "alice");
-      const createdV2 = await store.createBlob(ctx, blobV2);
-
-      assert.deepStrictEqual(createdV1.versionId, "");
-      assert.deepStrictEqual(createdV2.versionId, "");
-      assert.notDeepStrictEqual(createdV1.properties.contentLength, undefined);
-      assert.notDeepStrictEqual(createdV2.properties.contentLength, undefined);
-      assert.notDeepStrictEqual(
-        createdV1.properties.contentLength,
-        createdV2.properties.contentLength
-      );
-
-      // Download latest (no version) should return v2 (still versionId "")
-      const latest = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      assert.strictEqual(
-        latest.properties.contentLength,
-        blobV2.properties.contentLength
-      );
-      assert.strictEqual(
-        latest.versionId,
-        "",
-        "Still base version placeholder"
-      );
-    });
-
-    it("can retrieve a version created while versioning was enabled after disabling versioning @loki", async () => {
-      // Close the in-memory disabled store from beforeEach; we need persistence for this scenario
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create persistent store with versioning enabled (inMemory=false)
-      let persistent = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await persistent.init();
-      await persistent.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const blobV = buildBlockBlob(ACCOUNT, containerName, name, "body");
-
-      const createdBlob = await persistent.createBlob(ctx, blobV);
-      const versionId = createdBlob.versionId;
-      assert.ok(!isNullOrWhitespace(versionId));
-      await persistent.close(); // Do NOT clean so data persists
-
-      // 2. Recreate store with versioning disabled using same DB file
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // 3. Attempt to fetch explicitly by the version id created earlier
-      const fetched = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        versionId
-      );
-      assert.ok(
-        !isNullOrWhitespace(fetched.versionId),
-        "Fetched version should have a non-empty versionId"
-      );
-      assert.deepStrictEqual(fetched.versionId, versionId);
-    });
-
-    it("should not create versions for subsequent blob modifications when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blobV1 = buildBlockBlob(ACCOUNT, containerName, name, "version1");
-      await store.createBlob(ctx, blobV1);
-
-      // Multiple overwrites should all result in versionId ""
-      const blobV2 = buildBlockBlob(ACCOUNT, containerName, name, "version2");
-      await store.createBlob(ctx, blobV2);
-
-      const blobV3 = buildBlockBlob(ACCOUNT, containerName, name, "version3");
-      await store.createBlob(ctx, blobV3);
-
-      const latest = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      assert.strictEqual(latest.versionId, "");
-      assert.strictEqual(
-        latest.properties.contentLength,
-        blobV3.properties.contentLength
-      );
-    });
-
-    it("should handle delete operations without creating versions when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      // Delete the blob
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
-
-      // Verify blob is deleted
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          undefined,
-          undefined
-        );
-        assert.fail("Should have thrown error for deleted blob");
-      } catch (error) {
-        // Expected behavior - blob should be deleted
-      }
-    });
-
-    it("should allow creation of new blob with same name after deletion when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob1 = buildBlockBlob(ACCOUNT, containerName, name, "content1");
-      await store.createBlob(ctx, blob1);
-
-      // Delete the blob
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
-
-      // Create new blob with same name - should work and have versionId ""
-      const blob2 = buildBlockBlob(ACCOUNT, containerName, name, "content2");
-      const created = await store.createBlob(ctx, blob2);
-
-      assert.strictEqual(created.versionId, "");
-
-      const fetched = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      assert.strictEqual(fetched.versionId, "");
-      assert.strictEqual(
-        fetched.properties.contentLength,
-        blob2.properties.contentLength
-      );
-    });
-
-    // ================== SNAPSHOT TESTS WITH VERSIONING DISABLED ==================
-    it("should create snapshots without versions when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      const beforeSnapshot = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Take snapshot should not create version when versioning disabled
-      ctx.startTime = new Date(Date.now() + 100);
-      const snapshotResponse = await store.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-
-      assert.ok(snapshotResponse.snapshot);
-      assert.strictEqual(snapshotResponse.versionId, "");
-
-      // Current blob should still exist and not have a version
-      const afterSnapshot = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      assert.strictEqual(afterSnapshot.versionId, "");
-      assert.strictEqual(afterSnapshot.versionId, beforeSnapshot.versionId);
-    });
-
-    // ================== HTTP HEADERS TESTS WITH VERSIONING DISABLED ==================
-    it("should update HTTP headers in place without creating versions when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      const beforeHeaders = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.setBlobHTTPHeaders(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { blobContentType: "text/plain" }
-      );
-
-      const afterHeaders = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should update in place - same versionId (empty) but updated properties
-      assert.strictEqual(afterHeaders.versionId, beforeHeaders.versionId);
-      assert.strictEqual(afterHeaders.versionId, "");
-      assert.strictEqual(afterHeaders.properties.contentType, "text/plain");
-    });
-
-    // ================== BLOB TAGS TESTS WITH VERSIONING DISABLED ==================
-    it("should update blob tags in place without creating versions when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      const beforeTags = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.setBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined,
-        { blobTagSet: [{ key: "environment", value: "test" }] }
-      );
-
-      const afterTags = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should update in place - same versionId (empty)
-      assert.strictEqual(afterTags.versionId, beforeTags.versionId);
-      assert.strictEqual(afterTags.versionId, "");
-
-      // Verify tags are set
-      const tags = await store.getBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-      assert.deepStrictEqual(tags, {
-        blobTagSet: [{ key: "environment", value: "test" }]
-      });
-    });
-
-    // ================== TIER MANAGEMENT TESTS WITH VERSIONING DISABLED ==================
-    it("should update tier in place without creating versions when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      blob.properties.accessTier = Models.AccessTier.Hot;
-      await store.createBlob(ctx, blob);
-
-      const beforeTier = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Set tier should update in place
-      await store.setTier(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        Models.AccessTier.Cool,
-        undefined
-      );
-
-      const afterTier = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should update in place - same versionId (empty) but updated tier
-      assert.strictEqual(afterTier.versionId, beforeTier.versionId);
-      assert.strictEqual(afterTier.versionId, "");
-      assert.strictEqual(
-        afterTier.properties.accessTier,
-        Models.AccessTier.Cool
-      );
-    });
-
-    // ================== BLOB EXISTENCE AND PROPERTIES TESTS WITH VERSIONING DISABLED ==================
-    it("should check blob existence without version support when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      // Check existence should work
-      await store.checkBlobExist(ctx, ACCOUNT, containerName, name);
-
-      // Should throw for version-specific requests since versioning is disabled
-      try {
-        await store.checkBlobExist(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          "",
-          "2099-01-01T00:00:00.0000000Z"
-        );
-        assert.fail(
-          "Should have thrown for version-specific request when versioning disabled"
-        );
-      } catch (error) {
-        // Expected - version requests not supported when versioning disabled
-      }
-    });
-
-    it("should get properties without version support when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      // Set metadata
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { environment: "test" }
-      );
-
-      // Get properties should work
-      const props = await store.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-
-      assert.deepStrictEqual(props.metadata, { environment: "test" });
-    });
-
-    // ================== APPEND BLOB OPERATIONS TESTS WITH VERSIONING DISABLED ==================
-    it("should handle Append Block operations normally when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const appendBlob = buildAppendBlob(ACCOUNT, containerName, name);
-      await store.createBlob(ctx, appendBlob);
-
-      const afterCreate = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Append block
-      const block = {
-        accountName: ACCOUNT,
-        containerName,
-        blobName: name,
-        name: "append1",
-        size: 10,
-        persistency: { id: uuid(), offset: 0, count: 10 }
-      } as any;
-
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.appendBlock(ctx, block);
-
-      const afterAppend = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should update in place - same versionId (empty)
-      assert.strictEqual(afterAppend.versionId, afterCreate.versionId);
-      assert.strictEqual(afterAppend.versionId, "");
-      assert.strictEqual(afterAppend.properties.contentLength, 10);
-    });
-
-    // ================== PAGE BLOB OPERATIONS TESTS WITH VERSIONING DISABLED ==================
-    it("should handle Put Page operations normally when versioning disabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const pageBlob = buildPageBlob(ACCOUNT, containerName, name, 512);
-      await store.createBlob(ctx, pageBlob);
-
-      const afterCreate = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Put Page
-      const persistency = { id: uuid(), offset: 0, count: 512 };
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.uploadPages(ctx, pageBlob, 0, 511, persistency);
-
-      const afterUpload = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should update in place - same versionId (empty)
-      assert.strictEqual(afterUpload.versionId, afterCreate.versionId);
-      assert.strictEqual(afterUpload.versionId, "");
-    });
-
-    // ================== VERSION MODE TRANSITION TESTS (ENABLED → DISABLED) ==================
-    it("should handle setBlobMetadata correctly when disabling versioning after creating versions @loki", async () => {
-      // Close the in-memory disabled store from beforeEach; we need persistence for this scenario
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned blob
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      const createdBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
-
-      // Set metadata to create versions (should create version)
-      ctx.startTime = new Date(Date.now() + 100);
-      const modifiedMetadataBaseBlob = await enabledStore.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { versionedmeta: "value1" }
-      );
-      assert.ok(!isNullOrWhitespace(createdBaseBlob.versionId));
-      assert.ok(!isNullOrWhitespace(modifiedMetadataBaseBlob.versionId));
-      assert.notStrictEqual(
-        modifiedMetadataBaseBlob.versionId,
-        createdBaseBlob.versionId
-      );
-
-      const versionedFetched = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(!isNullOrWhitespace(versionedFetched.versionId));
-      assert.deepStrictEqual(versionedFetched.metadata, {
-        versionedmeta: "value1"
-      });
-      assert.strictEqual(
-        versionedFetched.versionId,
-        modifiedMetadataBaseBlob.versionId
-      );
-      const versionId = versionedFetched.versionId;
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Set metadata should NOT create new version (overwrite current)
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { disabledmeta: "value2" }
-      );
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should be same version (no new version created)
-      assert.strictEqual(current.versionId, "");
-      assert.notStrictEqual(current.versionId, versionId);
-      assert.deepStrictEqual(current.metadata, { disabledmeta: "value2" });
-
-      const firstVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        versionId
-      );
-      assert.strictEqual(firstVersion.versionId, versionId);
-    });
-
-    it("should handle setBlobHTTPHeaders correctly when disabling versioning after creating versions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned blob
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await enabledStore.createBlob(ctx, baseBlob);
-
-      // Set HTTP headers (should NOT create version even when versioning enabled)
-      ctx.startTime = new Date(Date.now() + 100);
-      await enabledStore.setBlobHTTPHeaders(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { blobContentType: "text/plain" }
-      );
-
-      const versionedFetched = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(!isNullOrWhitespace(versionedFetched.versionId));
-      assert.strictEqual(versionedFetched.properties.contentType, "text/plain");
-      const versionId = versionedFetched.versionId;
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Set headers should continue to NOT create version and update in place
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setBlobHTTPHeaders(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { blobContentType: "application/json" }
-      );
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should be same version (headers don't create versions in either mode)
-      assert.strictEqual(current.versionId, versionId);
-      assert.strictEqual(current.properties.contentType, "application/json");
-    });
-
-    it("should handle setBlobTag/getBlobTag correctly when disabling versioning after creating versions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned blob
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      const creaatedBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
-
-      // Set tags (should NOT create version even when versioning enabled)
-      ctx.startTime = new Date(Date.now() + 100);
-      await enabledStore.setBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined,
-        { blobTagSet: [{ key: "env", value: "test" }] }
-      );
-
-      const versionedFetched = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const versionId = versionedFetched.versionId;
-      assert.ok(!isNullOrWhitespace(versionId));
-      assert.strictEqual(creaatedBaseBlob.versionId, versionId);
-      const versionedTags = await enabledStore.getBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        versionId,
-        undefined
-      );
-      assert.deepStrictEqual(versionedTags, {
-        blobTagSet: [{ key: "env", value: "test" }]
-      });
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Set tags should continue to NOT create version and update in place
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined,
-        { blobTagSet: [{ key: "env", value: "prod" }] }
-      );
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should be same version (tags don't create versions in either mode)
-      assert.strictEqual(current.versionId, versionId);
-
-      const currentTags = await store.getBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        versionId,
-        undefined
-      );
-      assert.deepStrictEqual(currentTags, {
-        blobTagSet: [{ key: "env", value: "prod" }]
-      });
-    });
-
-    it("should handle setTier correctly when disabling versioning after creating versions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned blob
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      baseBlob.properties.accessTier = Models.AccessTier.Hot;
-      const blobCreated = await enabledStore.createBlob(ctx, baseBlob);
-
-      // Set tier (should NOT create version even when versioning enabled)
-      ctx.startTime = new Date(Date.now() + 100);
-      await enabledStore.setTier(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        Models.AccessTier.Cool,
-        undefined
-      );
-
-      const versionedFetched = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const versionId = versionedFetched.versionId;
-      assert.ok(!isNullOrWhitespace(versionId));
-      assert.strictEqual(blobCreated.versionId, versionId);
-      assert.strictEqual(
-        versionedFetched.properties.accessTier,
-        Models.AccessTier.Cool
-      );
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Set tier should continue to work and update in place
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setTier(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        Models.AccessTier.Archive,
-        undefined
-      );
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should be same version (tier operations don't create versions in either mode)
-      assert.strictEqual(current.versionId, versionId);
-      assert.strictEqual(
-        current.properties.accessTier,
-        Models.AccessTier.Archive
-      );
-    });
-
-    it("should handle checkBlobExist correctly when disabling versioning after creating versions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned blobs
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      const createdBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
-      const firstVersionId = createdBaseBlob.versionId;
-
-      // Create second version
-      ctx.startTime = new Date(Date.now() + 100);
-      const secondBlob = buildBlockBlob(ACCOUNT, containerName, name, "second");
-      const createdSecondBlob = await enabledStore.createBlob(ctx, secondBlob);
-
-      const current = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const currentVersionId = current.versionId;
-      assert.ok(!isNullOrWhitespace(currentVersionId));
-      assert.strictEqual(currentVersionId, createdSecondBlob.versionId);
-      assert.notStrictEqual(currentVersionId, firstVersionId);
-
-      // Get first version ID
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Check existence should work for current blob
-      await store.checkBlobExist(ctx, ACCOUNT, containerName, name);
-
-      // Should still be able to check existence by specific versionId
-      await store.checkBlobExist(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        "",
-        currentVersionId
-      );
-
-      // Previous versions should still be accessible by versionId
-      await store.checkBlobExist(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        "",
-        firstVersionId
-      );
-    });
-
-    it("should handle getBlobProperties correctly when disabling versioning after creating versions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned blobs
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await enabledStore.createBlob(ctx, baseBlob);
-
-      // Set metadata to create version
-      ctx.startTime = new Date(Date.now() + 100);
-      await enabledStore.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { env: "test" }
-      );
-
-      const secondVersion = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const secondVersionId = secondVersion.versionId;
-
-      // Create second version
-      ctx.startTime = new Date(Date.now() + 200);
-      await enabledStore.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { env: "prod" }
-      );
-
-      const current = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const currentVersionId = current.versionId;
-      assert.ok(!isNullOrWhitespace(currentVersionId));
-      assert.notStrictEqual(currentVersionId, secondVersionId);
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Get properties should work for current version
-      const currentProps = await store.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-      assert.deepStrictEqual(currentProps.metadata, { env: "prod" });
-
-      // Should still be able to get properties for specific versions by versionId
-      const currentPropsByVersion = await store.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        currentVersionId,
-        undefined
-      );
-      assert.deepStrictEqual(
-        currentPropsByVersion.metadata,
-        currentProps.metadata
-      );
-
-      const secondVersionProps = await store.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        secondVersionId,
-        undefined
-      );
-      assert.deepStrictEqual(secondVersionProps.metadata, { env: "test" });
-    });
-
-    it("should handle createSnapshot correctly when disabling versioning after creating versions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned blob
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      const createdBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
-
-      // Create snapshot (should create new version when versioning enabled)
-      ctx.startTime = new Date(Date.now() + 100);
-      const snapshotResponse1 = await enabledStore.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-      assert.ok(snapshotResponse1.snapshot);
-      assert.ok(!isNullOrWhitespace(snapshotResponse1.versionId));
-      assert.notStrictEqual(
-        snapshotResponse1.versionId,
-        createdBaseBlob.versionId
-      );
-
-      const versionedFetched = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const versionId = versionedFetched.versionId;
-      assert.ok(!isNullOrWhitespace(versionId));
-      assert.strictEqual(versionId, snapshotResponse1.versionId);
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Create snapshot should NOT create new version when versioning disabled
-      ctx.startTime = new Date(Date.now() + 200);
-      const snapshotResponse2 = await store.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-      assert.ok(snapshotResponse2.snapshot);
-      assert.strictEqual(snapshotResponse2.versionId, "");
-
-      try {
-        // Snapshotting acts as a "write"
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          undefined,
-          undefined
-        );
-        assert.fail("Expected error to be thrown");
-      } catch (error) {
-        assert.ok(error);
-      }
-    });
-
-    it("should handle appendBlock correctly when disabling versioning after creating versions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned append blob
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseAppendBlob = buildAppendBlob(ACCOUNT, containerName, name);
-      await enabledStore.createBlob(ctx, baseAppendBlob);
-
-      // Append block (should NOT create version even when versioning enabled)
-      const block1 = {
-        accountName: ACCOUNT,
-        containerName,
-        blobName: name,
-        name: "append1",
-        size: 10,
-        persistency: { id: uuid(), offset: 0, count: 10 }
-      } as any;
-
-      ctx.startTime = new Date(Date.now() + 100);
-      await enabledStore.appendBlock(ctx, block1);
-
-      const versionedFetched = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(!isNullOrWhitespace(versionedFetched.versionId));
-      assert.strictEqual(versionedFetched.properties.contentLength, 10);
-      const versionId = versionedFetched.versionId;
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Append block should continue to NOT create version and update in place
-      const block2 = {
-        accountName: ACCOUNT,
-        containerName,
-        blobName: name,
-        name: "append2",
-        size: 15,
-        persistency: { id: uuid(), offset: 10, count: 15 }
-      } as any;
-
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.appendBlock(ctx, block2);
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should be same version (append operations don't create versions in either mode)
-      assert.strictEqual(current.versionId, versionId);
-      assert.strictEqual(current.properties.contentLength, 25);
-    });
-
-    it("should handle uploadPages correctly when disabling versioning after creating versions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned page blob
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const basePageBlob = buildPageBlob(ACCOUNT, containerName, name, 512);
-      await enabledStore.createBlob(ctx, basePageBlob);
-
-      // Upload pages (should NOT create version even when versioning enabled)
-      const persistency1 = { id: uuid(), offset: 0, count: 512 };
-      ctx.startTime = new Date(Date.now() + 100);
-      await enabledStore.uploadPages(ctx, basePageBlob, 0, 511, persistency1);
-
-      const versionedFetched = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(!isNullOrWhitespace(versionedFetched.versionId));
-      const versionId = versionedFetched.versionId;
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Upload pages should continue to NOT create version and update in place
-      const persistency2 = { id: uuid(), offset: 0, count: 512 };
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.uploadPages(ctx, basePageBlob, 0, 511, persistency2);
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should be same version (page operations don't create versions in either mode)
-      assert.strictEqual(current.versionId, versionId);
-    });
-
-    it("should handle deleteBlob correctly when disabling versioning after creating versions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create versioned blobs
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      const createdBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
-
-      // Create second version
-      ctx.startTime = new Date(Date.now() + 100);
-      const secondBlob = buildBlockBlob(ACCOUNT, containerName, name, "second");
-      await enabledStore.createBlob(ctx, secondBlob);
-
-      const beforeDelete = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const currentVersionId = beforeDelete.versionId;
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // Delete current blob should completely remove it (not make it a previous version)
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
-
-      // Current version should no longer exist
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          undefined,
-          undefined
-        );
-        assert.fail("Should have thrown error for deleted current blob");
-      } catch (error) {
-        // Expected
-      }
-
-      // But should still be able to access previous versions by specific versionId
-      const deletedVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        createdBaseBlob.versionId
-      );
-      assert.ok(!isNullOrWhitespace(deletedVersion.versionId));
-      assert.notStrictEqual(deletedVersion.versionId, currentVersionId);
-      assert.strictEqual(deletedVersion.versionId, createdBaseBlob.versionId);
-
-      // Should be able to delete specific version by versionId
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-        versionId: createdBaseBlob.versionId
-      });
-
-      // That specific version should no longer exist
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          undefined,
-          createdBaseBlob.versionId
-        );
-        assert.fail("Should have thrown error for deleted specific version");
-      } catch (error) {
-        // Expected
-      }
-    });
-
-    it("should preserve existing versions and allow operations on them when versioning is disabled @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning ENABLED and create multiple versions
-      let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await enabledStore.init();
-      await enabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-
-      // Create first version
-      const blob1 = buildBlockBlob(ACCOUNT, containerName, name, "version1");
-      await enabledStore.createBlob(ctx, blob1);
-      const version1 = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const version1Id = version1.versionId;
-
-      // Create second version
-      ctx.startTime = new Date(Date.now() + 100);
-      const blob2 = buildBlockBlob(ACCOUNT, containerName, name, "version2");
-      await enabledStore.createBlob(ctx, blob2);
-      const version2 = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const version2Id = version2.versionId;
-
-      // Create third version
-      ctx.startTime = new Date(Date.now() + 200);
-      const blob3 = buildBlockBlob(ACCOUNT, containerName, name, "version3");
-      await enabledStore.createBlob(ctx, blob3);
-      const version3 = await enabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const version3Id = version3.versionId;
-      await enabledStore.close();
-
-      // 2. Re-open with versioning DISABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await store.init();
-
-      // All existing versions should remain accessible by versionId
-      const fetchedV1 = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        version1Id
-      );
-      assert.strictEqual(
-        fetchedV1.properties.contentLength,
-        blob1.properties.contentLength
-      );
-      assert.strictEqual(fetchedV1.versionId, version1Id);
-
-      const fetchedV2 = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        version2Id
-      );
-      assert.strictEqual(
-        fetchedV2.properties.contentLength,
-        blob2.properties.contentLength
-      );
-      assert.strictEqual(fetchedV2.versionId, version2Id);
-
-      const fetchedV3 = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        version3Id
-      );
-      assert.strictEqual(
-        fetchedV3.properties.contentLength,
-        blob3.properties.contentLength
-      );
-      assert.strictEqual(fetchedV3.versionId, version3Id);
-
-      // Current version should be the latest (version3)
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(current.versionId, version3Id);
-      assert.strictEqual(
-        current.properties.contentLength,
-        blob3.properties.contentLength
-      );
-
-      // Modifying current should create new version
-      ctx.startTime = new Date(Date.now() + 300);
-      const newBlob = buildBlockBlob(
-        ACCOUNT,
-        containerName,
-        name,
-        "modified_no_version"
-      );
-      await store.createBlob(ctx, newBlob);
-
-      const afterModify = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should change to empty string
-      assert.strictEqual(afterModify.versionId, "");
-      assert.notStrictEqual(afterModify.versionId, version3Id);
-
-      // Previous versions should still exist and be unchanged
-      const stillV1 = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        version1Id
-      );
-      assert.strictEqual(
-        stillV1.properties.contentLength,
-        blob1.properties.contentLength
-      );
-
-      const stillV2 = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        version2Id
-      );
-      assert.strictEqual(
-        stillV2.properties.contentLength,
-        blob2.properties.contentLength
-      );
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Set tags should continue to NOT create version and update in place
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined,
+      { blobTagSet: [{ key: "env", value: "prod" }] }
+    );
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should be same version (tags don't create versions in either mode)
+    assert.strictEqual(current.versionId, versionId);
+
+    const currentTags = await store.getBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      versionId,
+      undefined
+    );
+    assert.deepStrictEqual(currentTags, {
+      blobTagSet: [{ key: "env", value: "prod" }]
     });
   });
 
-  describe("When blob versioning enabled", () => {
-    let store: LokiBlobMetadataStore;
-    let containerName: string;
-    let ctx: Context;
-    const DB_FILE = "__test_db_blob__.json"; // standard shared test db path
+  it("should handle setTier correctly when disabling versioning after creating versions @loki", async () => {
+    await store.close();
+    await store.clean();
 
-    beforeEach(async () => {
-      ctx = createContext();
-      containerName = `container-${uuid()}`;
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-      await store.createContainer(ctx, buildContainer(ACCOUNT, containerName));
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned blob
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    baseBlob.properties.accessTier = Models.AccessTier.Hot;
+    const blobCreated = await enabledStore.createBlob(ctx, baseBlob);
+
+    // Set tier (should NOT create version even when versioning enabled)
+    ctx.startTime = new Date(Date.now() + 100);
+    await enabledStore.setTier(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      Models.AccessTier.Cool,
+      undefined
+    );
+
+    const versionedFetched = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const versionId = versionedFetched.versionId;
+    assert.ok(!isNullOrWhitespace(versionId));
+    assert.strictEqual(blobCreated.versionId, versionId);
+    assert.strictEqual(
+      versionedFetched.properties.accessTier,
+      Models.AccessTier.Cool
+    );
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Set tier should continue to work and update in place
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setTier(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      Models.AccessTier.Archive,
+      undefined
+    );
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should be same version (tier operations don't create versions in either mode)
+    assert.strictEqual(current.versionId, versionId);
+    assert.strictEqual(
+      current.properties.accessTier,
+      Models.AccessTier.Archive
+    );
+  });
+
+  it("should handle checkBlobExist correctly when disabling versioning after creating versions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned blobs
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    const createdBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
+    const firstVersionId = createdBaseBlob.versionId;
+
+    // Create second version
+    ctx.startTime = new Date(Date.now() + 100);
+    const secondBlob = buildBlockBlob(ACCOUNT, containerName, name, "second");
+    const createdSecondBlob = await enabledStore.createBlob(ctx, secondBlob);
+
+    const current = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const currentVersionId = current.versionId;
+    assert.ok(!isNullOrWhitespace(currentVersionId));
+    assert.strictEqual(currentVersionId, createdSecondBlob.versionId);
+    assert.notStrictEqual(currentVersionId, firstVersionId);
+
+    // Get first version ID
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Check existence should work for current blob
+    await store.checkBlobExist(ctx, ACCOUNT, containerName, name);
+
+    // Should still be able to check existence by specific versionId
+    await store.checkBlobExist(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      "",
+      currentVersionId
+    );
+
+    // Previous versions should still be accessible by versionId
+    await store.checkBlobExist(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      "",
+      firstVersionId
+    );
+  });
+
+  it("should handle getBlobProperties correctly when disabling versioning after creating versions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned blobs
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await enabledStore.createBlob(ctx, baseBlob);
+
+    // Set metadata to create version
+    ctx.startTime = new Date(Date.now() + 100);
+    await enabledStore.setBlobMetadata(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { env: "test" }
+    );
+
+    const secondVersion = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const secondVersionId = secondVersion.versionId;
+
+    // Create second version
+    ctx.startTime = new Date(Date.now() + 200);
+    await enabledStore.setBlobMetadata(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { env: "prod" }
+    );
+
+    const current = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const currentVersionId = current.versionId;
+    assert.ok(!isNullOrWhitespace(currentVersionId));
+    assert.notStrictEqual(currentVersionId, secondVersionId);
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Get properties should work for current version
+    const currentProps = await store.getBlobProperties(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined
+    );
+    assert.deepStrictEqual(currentProps.metadata, { env: "prod" });
+
+    // Should still be able to get properties for specific versions by versionId
+    const currentPropsByVersion = await store.getBlobProperties(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      currentVersionId,
+      undefined
+    );
+    assert.deepStrictEqual(
+      currentPropsByVersion.metadata,
+      currentProps.metadata
+    );
+
+    const secondVersionProps = await store.getBlobProperties(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      secondVersionId,
+      undefined
+    );
+    assert.deepStrictEqual(secondVersionProps.metadata, { env: "test" });
+  });
+
+  it("should handle createSnapshot correctly when disabling versioning after creating versions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned blob
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    const createdBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
+
+    // Create snapshot (should create new version when versioning enabled)
+    ctx.startTime = new Date(Date.now() + 100);
+    const snapshotResponse1 = await enabledStore.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    assert.ok(snapshotResponse1.snapshot);
+    assert.ok(!isNullOrWhitespace(snapshotResponse1.versionId));
+    assert.notStrictEqual(
+      snapshotResponse1.versionId,
+      createdBaseBlob.versionId
+    );
+
+    const versionedFetched = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const versionId = versionedFetched.versionId;
+    assert.ok(!isNullOrWhitespace(versionId));
+    assert.strictEqual(versionId, snapshotResponse1.versionId);
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Create snapshot should NOT create new version when versioning disabled
+    ctx.startTime = new Date(Date.now() + 200);
+    const snapshotResponse2 = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    assert.ok(snapshotResponse2.snapshot);
+    assert.strictEqual(snapshotResponse2.versionId, "");
+
+    try {
+      // Snapshotting acts as a "write"
+      await store.downloadBlob(
+        ctx,
+        ACCOUNT,
+        containerName,
+        name,
+        undefined,
+        undefined
+      );
+      assert.fail("Expected error to be thrown");
+    } catch (error) {
+      assert.ok(error);
+    }
+  });
+
+  it("should handle appendBlock correctly when disabling versioning after creating versions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned append blob
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseAppendBlob = buildAppendBlob(ACCOUNT, containerName, name);
+    await enabledStore.createBlob(ctx, baseAppendBlob);
+
+    // Append block (should NOT create version even when versioning enabled)
+    const block1 = {
+      accountName: ACCOUNT,
+      containerName,
+      blobName: name,
+      name: "append1",
+      size: 10,
+      persistency: { id: uuid(), offset: 0, count: 10 }
+    } as any;
+
+    ctx.startTime = new Date(Date.now() + 100);
+    await enabledStore.appendBlock(ctx, block1);
+
+    const versionedFetched = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(!isNullOrWhitespace(versionedFetched.versionId));
+    assert.strictEqual(versionedFetched.properties.contentLength, 10);
+    const versionId = versionedFetched.versionId;
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Append block should continue to NOT create version and update in place
+    const block2 = {
+      accountName: ACCOUNT,
+      containerName,
+      blobName: name,
+      name: "append2",
+      size: 15,
+      persistency: { id: uuid(), offset: 10, count: 15 }
+    } as any;
+
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.appendBlock(ctx, block2);
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should be same version (append operations don't create versions in either mode)
+    assert.strictEqual(current.versionId, versionId);
+    assert.strictEqual(current.properties.contentLength, 25);
+  });
+
+  it("should handle uploadPages correctly when disabling versioning after creating versions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned page blob
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const basePageBlob = buildPageBlob(ACCOUNT, containerName, name, 512);
+    await enabledStore.createBlob(ctx, basePageBlob);
+
+    // Upload pages (should NOT create version even when versioning enabled)
+    const persistency1 = { id: uuid(), offset: 0, count: 512 };
+    ctx.startTime = new Date(Date.now() + 100);
+    await enabledStore.uploadPages(ctx, basePageBlob, 0, 511, persistency1);
+
+    const versionedFetched = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(!isNullOrWhitespace(versionedFetched.versionId));
+    const versionId = versionedFetched.versionId;
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Upload pages should continue to NOT create version and update in place
+    const persistency2 = { id: uuid(), offset: 0, count: 512 };
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.uploadPages(ctx, basePageBlob, 0, 511, persistency2);
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should be same version (page operations don't create versions in either mode)
+    assert.strictEqual(current.versionId, versionId);
+  });
+
+  it("should handle deleteBlob correctly when disabling versioning after creating versions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create versioned blobs
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    const createdBaseBlob = await enabledStore.createBlob(ctx, baseBlob);
+
+    // Create second version
+    ctx.startTime = new Date(Date.now() + 100);
+    const secondBlob = buildBlockBlob(ACCOUNT, containerName, name, "second");
+    await enabledStore.createBlob(ctx, secondBlob);
+
+    const beforeDelete = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const currentVersionId = beforeDelete.versionId;
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // Delete current blob should completely remove it (not make it a previous version)
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
+
+    // Current version should no longer exist
+    try {
+      await store.downloadBlob(
+        ctx,
+        ACCOUNT,
+        containerName,
+        name,
+        undefined,
+        undefined
+      );
+      assert.fail("Should have thrown error for deleted current blob");
+    } catch (error) {
+      // Expected
+    }
+
+    // But should still be able to access previous versions by specific versionId
+    const deletedVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      createdBaseBlob.versionId
+    );
+    assert.ok(!isNullOrWhitespace(deletedVersion.versionId));
+    assert.notStrictEqual(deletedVersion.versionId, currentVersionId);
+    assert.strictEqual(deletedVersion.versionId, createdBaseBlob.versionId);
+
+    // Should be able to delete specific version by versionId
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
+      versionId: createdBaseBlob.versionId
     });
 
-    afterEach(async () => {
-      await store.close();
-      await store.clean();
+    // That specific version should no longer exist
+    try {
+      await store.downloadBlob(
+        ctx,
+        ACCOUNT,
+        containerName,
+        name,
+        undefined,
+        createdBaseBlob.versionId
+      );
+      assert.fail("Should have thrown error for deleted specific version");
+    } catch (error) {
+      // Expected
+    }
+  });
+
+  it("should preserve existing versions and allow operations on them when versioning is disabled @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning ENABLED and create multiple versions
+    let enabledStore = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await enabledStore.init();
+    await enabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+
+    // Create first version
+    const blob1 = buildBlockBlob(ACCOUNT, containerName, name, "version1");
+    await enabledStore.createBlob(ctx, blob1);
+    const version1 = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const version1Id = version1.versionId;
+
+    // Create second version
+    ctx.startTime = new Date(Date.now() + 100);
+    const blob2 = buildBlockBlob(ACCOUNT, containerName, name, "version2");
+    await enabledStore.createBlob(ctx, blob2);
+    const version2 = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const version2Id = version2.versionId;
+
+    // Create third version
+    ctx.startTime = new Date(Date.now() + 200);
+    const blob3 = buildBlockBlob(ACCOUNT, containerName, name, "version3");
+    await enabledStore.createBlob(ctx, blob3);
+    const version3 = await enabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const version3Id = version3.versionId;
+    await enabledStore.close();
+
+    // 2. Re-open with versioning DISABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await store.init();
+
+    // All existing versions should remain accessible by versionId
+    const fetchedV1 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      version1Id
+    );
+    assert.strictEqual(
+      fetchedV1.properties.contentLength,
+      blob1.properties.contentLength
+    );
+    assert.strictEqual(fetchedV1.versionId, version1Id);
+
+    const fetchedV2 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      version2Id
+    );
+    assert.strictEqual(
+      fetchedV2.properties.contentLength,
+      blob2.properties.contentLength
+    );
+    assert.strictEqual(fetchedV2.versionId, version2Id);
+
+    const fetchedV3 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      version3Id
+    );
+    assert.strictEqual(
+      fetchedV3.properties.contentLength,
+      blob3.properties.contentLength
+    );
+    assert.strictEqual(fetchedV3.versionId, version3Id);
+
+    // Current version should be the latest (version3)
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(current.versionId, version3Id);
+    assert.strictEqual(
+      current.properties.contentLength,
+      blob3.properties.contentLength
+    );
+
+    // Modifying current should create new version
+    ctx.startTime = new Date(Date.now() + 300);
+    const newBlob = buildBlockBlob(
+      ACCOUNT,
+      containerName,
+      name,
+      "modified_no_version"
+    );
+    await store.createBlob(ctx, newBlob);
+
+    const afterModify = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should change to empty string
+    assert.strictEqual(afterModify.versionId, "");
+    assert.notStrictEqual(afterModify.versionId, version3Id);
+
+    // Previous versions should still exist and be unchanged
+    const stillV1 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      version1Id
+    );
+    assert.strictEqual(
+      stillV1.properties.contentLength,
+      blob1.properties.contentLength
+    );
+
+    const stillV2 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      version2Id
+    );
+    assert.strictEqual(
+      stillV2.properties.contentLength,
+      blob2.properties.contentLength
+    );
+  });
+
+  it("creates a new version for each blob creation and marks previous current as not current @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "v1");
+    await store.createBlob(ctx, v1);
+    const afterV1 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(
+      afterV1.versionId || afterV1.versionId === "",
+      "First creation should have a version id (may be empty transitioning)"
+    );
+
+    // Second create -> new version id expected
+    ctx.startTime = new Date(Date.now() + 10); // ensure different timestamp base
+    const v2 = buildBlockBlob(ACCOUNT, containerName, name, "v2");
+    await store.createBlob(ctx, v2);
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.notStrictEqual(
+      current.properties.etag,
+      afterV1.properties.etag,
+      "Etag should change for new version"
+    );
+    assert.ok(current.isCurrentVersion, "Latest should be current version");
+  });
+
+  it("promotes a non-versioned base blob (created with versioning disabled) to have a versionId equal to its original lastModified when versioning is later enabled @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED (persistent) and create base blob (versionId will be "").
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await disabledStore.createBlob(ctx, baseBlob);
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(
+      baseFetched.versionId,
+      "",
+      "Pre-versioning blob should have empty versionId"
+    );
+    const originalLastModifiedIso =
+      baseFetched.properties.lastModified.toISOString();
+    await disabledStore.close();
+
+    // 2. Re-open SAME DB with versioning ENABLED.
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // 3. Create a new version (same name). This should assign a versionId to prior base blob
+    //    using its lastModified timestamp, and mark it as not current.
+    ctx.startTime = new Date(Date.now() + 200); // ensure different timestamp for new current version
+    const secondBlob = buildBlockBlob(ACCOUNT, containerName, name, "second");
+    const newBlobVer = await store.createBlob(ctx, secondBlob);
+    assert.ok(
+      !isNullOrWhitespace(newBlobVer.versionId),
+      "New blob version should have a versionId"
+    );
+
+    // 4. Fetch current (no version) and previous (by derived versionId)
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(
+      current.isCurrentVersion,
+      "Latest blob should be current after enabling versioning"
+    );
+    assert.ok(
+      !isNullOrWhitespace(current.versionId),
+      "Current blob should now have a non-empty versionId"
+    );
+
+    const previous = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      originalLastModifiedIso
+    );
+    assert.strictEqual(
+      previous.versionId,
+      originalLastModifiedIso,
+      "Previous base blob should be promoted with versionId equal to its original lastModified ISO string"
+    );
+    assert.strictEqual(
+      previous.isCurrentVersion,
+      false,
+      "Previous version should no longer be current"
+    );
+    assert.notStrictEqual(
+      previous.versionId,
+      current.versionId,
+      "Current versionId should differ from promoted previous versionId"
+    );
+  });
+
+  it("allows addressing previous version by its versionId @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "v1");
+    await store.createBlob(ctx, v1);
+    const first = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    ctx.startTime = new Date(Date.now() + 100);
+    const v2 = buildBlockBlob(ACCOUNT, containerName, name, "v2");
+    await store.createBlob(ctx, v2);
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Try to fetch previous by versionId - should have non-empty version ID
+    assert.ok(!isNullOrWhitespace(first.versionId));
+    const previousFetched = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      first.versionId
+    );
+    assert.ok(previousFetched.versionId === first.versionId);
+    assert.ok(current.isCurrentVersion);
+  });
+
+  it("should assign unique version IDs based on timestamp when creating versions @loki", async () => {
+    const name = `blob-${uuid()}`;
+
+    // Create first version
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "v1");
+    const created1 = await store.createBlob(ctx, v1);
+
+    // Wait a moment to ensure different timestamp
+    ctx.startTime = new Date(Date.now() + 100);
+
+    // Create second version
+    const v2 = buildBlockBlob(ACCOUNT, containerName, name, "v2");
+    const created2 = await store.createBlob(ctx, v2);
+
+    // Version IDs should be different
+    assert.notStrictEqual(created1.versionId, created2.versionId);
+    assert.ok(!isNullOrWhitespace(created1.versionId));
+    assert.ok(!isNullOrWhitespace(created2.versionId));
+  });
+
+  it("should maintain previous versions when creating new versions @loki", async () => {
+    const name = `blob-${uuid()}`;
+
+    // Create first version
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "ver1");
+    const created1 = await store.createBlob(ctx, v1);
+    const version1Id = created1.versionId;
+
+    // Create second version
+    ctx.startTime = new Date(Date.now() + 100);
+    const v2 = buildBlockBlob(ACCOUNT, containerName, name, "ver22");
+    const created2 = await store.createBlob(ctx, v2);
+    const version2Id = created2.versionId;
+
+    // Create third version
+    ctx.startTime = new Date(Date.now() + 200);
+    const v3 = buildBlockBlob(ACCOUNT, containerName, name, "ver333");
+    await store.createBlob(ctx, v3);
+
+    // Current should be v3
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(
+      current.properties.contentLength,
+      v3.properties.contentLength
+    );
+    assert.ok(current.isCurrentVersion);
+
+    // Previous versions should be accessible by versionId
+    assert.ok(!isNullOrWhitespace(version1Id));
+    const prev1 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      version1Id
+    );
+    assert.strictEqual(
+      prev1.properties.contentLength,
+      v1.properties.contentLength
+    );
+    assert.strictEqual(prev1.isCurrentVersion, false);
+
+    assert.ok(!isNullOrWhitespace(version2Id));
+    const prev2 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      version2Id
+    );
+    assert.strictEqual(
+      prev2.properties.contentLength,
+      v2.properties.contentLength
+    );
+    assert.strictEqual(prev2.isCurrentVersion, false);
+  });
+
+  it("should handle delete operations by making current version a previous version @loki", async () => {
+    const name = `blob-${uuid()}`;
+
+    // Create version
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, v1);
+
+    // Verify blob exists and is current
+    const beforeDelete = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(beforeDelete.isCurrentVersion);
+    const versionIdBeforeDelete = beforeDelete.versionId;
+
+    // Delete the blob (without version ID = delete current)
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
+
+    // Current version should no longer exist
+    try {
+      await store.downloadBlob(
+        ctx,
+        ACCOUNT,
+        containerName,
+        name,
+        undefined,
+        undefined
+      );
+      assert.fail("Should have thrown error for deleted current blob");
+    } catch (error) {
+      // Expected - no current version after delete
+    }
+
+    // Previous version should still be accessible by version ID
+    assert.ok(!isNullOrWhitespace(versionIdBeforeDelete));
+    const previousVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      versionIdBeforeDelete
+    );
+    assert.strictEqual(previousVersion.isCurrentVersion, false);
+    assert.strictEqual(previousVersion.versionId, versionIdBeforeDelete);
+  });
+
+  it("should allow creating new current version after deletion @loki", async () => {
+    const name = `blob-${uuid()}`;
+
+    // Create and delete a version
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "content1");
+    await store.createBlob(ctx, v1);
+    const beforeDelete = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    const deletedVersionId = beforeDelete.versionId;
+
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
+
+    // Create new blob with same name
+    ctx.startTime = new Date(Date.now() + 100);
+    const v2 = buildBlockBlob(ACCOUNT, containerName, name, "content2");
+    await store.createBlob(ctx, v2);
+
+    // New blob should be current version
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(current.isCurrentVersion);
+    assert.strictEqual(
+      current.properties.contentLength,
+      v2.properties.contentLength
+    );
+    assert.notStrictEqual(current.versionId, deletedVersionId);
+
+    // Previous version should still exist as non-current
+    assert.ok(!isNullOrWhitespace(deletedVersionId));
+    const previous = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      deletedVersionId
+    );
+    assert.strictEqual(previous.isCurrentVersion, false);
+  });
+
+  it("should allow deleting specific versions by versionId @loki", async () => {
+    const name = `blob-${uuid()}`;
+
+    // Create multiple versions
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "version1");
+    const created1 = await store.createBlob(ctx, v1);
+    const version1Id = created1.versionId;
+
+    ctx.startTime = new Date(Date.now() + 100);
+    const v2 = buildBlockBlob(ACCOUNT, containerName, name, "version2");
+    await store.createBlob(ctx, v2);
+
+    // Delete specific version (v1) by versionId
+    assert.ok(!isNullOrWhitespace(version1Id));
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
+      versionId: version1Id
     });
 
-    it("creates a new version for each blob creation and marks previous current as not current @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "v1");
-      await store.createBlob(ctx, v1);
-      const afterV1 = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(
-        afterV1.versionId || afterV1.versionId === "",
-        "First creation should have a version id (may be empty transitioning)"
-      );
+    // Current version (v2) should still exist
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(current.isCurrentVersion);
+    assert.strictEqual(
+      current.properties.contentLength,
+      v2.properties.contentLength
+    );
 
-      // Second create -> new version id expected
-      ctx.startTime = new Date(Date.now() + 10); // ensure different timestamp base
-      const v2 = buildBlockBlob(ACCOUNT, containerName, name, "v2");
-      await store.createBlob(ctx, v2);
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.notStrictEqual(
-        current.properties.etag,
-        afterV1.properties.etag,
-        "Etag should change for new version"
-      );
-      assert.ok(current.isCurrentVersion, "Latest should be current version");
-    });
-
-    it("promotes a non-versioned base blob (created with versioning disabled) to have a versionId equal to its original lastModified when versioning is later enabled @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED (persistent) and create base blob (versionId will be "").
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await disabledStore.createBlob(ctx, baseBlob);
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(
-        baseFetched.versionId,
-        "",
-        "Pre-versioning blob should have empty versionId"
-      );
-      const originalLastModifiedIso =
-        baseFetched.properties.lastModified.toISOString();
-      await disabledStore.close();
-
-      // 2. Re-open SAME DB with versioning ENABLED.
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // 3. Create a new version (same name). This should assign a versionId to prior base blob
-      //    using its lastModified timestamp, and mark it as not current.
-      ctx.startTime = new Date(Date.now() + 200); // ensure different timestamp for new current version
-      const secondBlob = buildBlockBlob(ACCOUNT, containerName, name, "second");
-      const newBlobVer = await store.createBlob(ctx, secondBlob);
-      assert.ok(
-        !isNullOrWhitespace(newBlobVer.versionId),
-        "New blob version should have a versionId"
-      );
-
-      // 4. Fetch current (no version) and previous (by derived versionId)
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(
-        current.isCurrentVersion,
-        "Latest blob should be current after enabling versioning"
-      );
-      assert.ok(
-        !isNullOrWhitespace(current.versionId),
-        "Current blob should now have a non-empty versionId"
-      );
-
-      const previous = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        originalLastModifiedIso
-      );
-      assert.strictEqual(
-        previous.versionId,
-        originalLastModifiedIso,
-        "Previous base blob should be promoted with versionId equal to its original lastModified ISO string"
-      );
-      assert.strictEqual(
-        previous.isCurrentVersion,
-        false,
-        "Previous version should no longer be current"
-      );
-      assert.notStrictEqual(
-        previous.versionId,
-        current.versionId,
-        "Current versionId should differ from promoted previous versionId"
-      );
-    });
-
-    it("allows addressing previous version by its versionId @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "v1");
-      await store.createBlob(ctx, v1);
-      const first = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      ctx.startTime = new Date(Date.now() + 100);
-      const v2 = buildBlockBlob(ACCOUNT, containerName, name, "v2");
-      await store.createBlob(ctx, v2);
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Try to fetch previous by versionId - should have non-empty version ID
-      assert.ok(!isNullOrWhitespace(first.versionId));
-      const previousFetched = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        first.versionId
-      );
-      assert.ok(previousFetched.versionId === first.versionId);
-      assert.ok(current.isCurrentVersion);
-    });
-
-    it("should assign unique version IDs based on timestamp when creating versions @loki", async () => {
-      const name = `blob-${uuid()}`;
-
-      // Create first version
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "v1");
-      const created1 = await store.createBlob(ctx, v1);
-
-      // Wait a moment to ensure different timestamp
-      ctx.startTime = new Date(Date.now() + 100);
-
-      // Create second version
-      const v2 = buildBlockBlob(ACCOUNT, containerName, name, "v2");
-      const created2 = await store.createBlob(ctx, v2);
-
-      // Version IDs should be different
-      assert.notStrictEqual(created1.versionId, created2.versionId);
-      assert.ok(!isNullOrWhitespace(created1.versionId));
-      assert.ok(!isNullOrWhitespace(created2.versionId));
-    });
-
-    it("should maintain previous versions when creating new versions @loki", async () => {
-      const name = `blob-${uuid()}`;
-
-      // Create first version
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "ver1");
-      const created1 = await store.createBlob(ctx, v1);
-      const version1Id = created1.versionId;
-
-      // Create second version
-      ctx.startTime = new Date(Date.now() + 100);
-      const v2 = buildBlockBlob(ACCOUNT, containerName, name, "ver22");
-      const created2 = await store.createBlob(ctx, v2);
-      const version2Id = created2.versionId;
-
-      // Create third version
-      ctx.startTime = new Date(Date.now() + 200);
-      const v3 = buildBlockBlob(ACCOUNT, containerName, name, "ver333");
-      await store.createBlob(ctx, v3);
-
-      // Current should be v3
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(
-        current.properties.contentLength,
-        v3.properties.contentLength
-      );
-      assert.ok(current.isCurrentVersion);
-
-      // Previous versions should be accessible by versionId
-      assert.ok(!isNullOrWhitespace(version1Id));
-      const prev1 = await store.downloadBlob(
+    // Deleted version should no longer be accessible
+    try {
+      await store.downloadBlob(
         ctx,
         ACCOUNT,
         containerName,
@@ -1949,1935 +1392,1764 @@ describe("LokiBlobMetadataStoreVersioning", () => {
         undefined,
         version1Id
       );
-      assert.strictEqual(
-        prev1.properties.contentLength,
-        v1.properties.contentLength
-      );
-      assert.strictEqual(prev1.isCurrentVersion, false);
+      assert.fail("Should have thrown error for deleted version");
+    } catch (error) {
+      // Expected behavior
+    }
+  });
 
-      assert.ok(!isNullOrWhitespace(version2Id));
-      const prev2 = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        version2Id
-      );
-      assert.strictEqual(
-        prev2.properties.contentLength,
-        v2.properties.contentLength
-      );
-      assert.strictEqual(prev2.isCurrentVersion, false);
+  it("should create versions for write operations on existing blobs @loki", async () => {
+    const name = `blob-${uuid()}`;
+
+    // Create initial blob
+    const initialBlob = buildBlockBlob(ACCOUNT, containerName, name, "initial");
+    await store.createBlob(ctx, initialBlob);
+    const firstVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Update metadata (write operation) should create new version
+    ctx.startTime = new Date(Date.now() + 100);
+    await store.setBlobMetadata(ctx, ACCOUNT, containerName, name, undefined, {
+      custommeta: "value"
     });
 
-    it("should handle delete operations by making current version a previous version @loki", async () => {
-      const name = `blob-${uuid()}`;
-
-      // Create version
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, v1);
-
-      // Verify blob exists and is current
-      const beforeDelete = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(beforeDelete.isCurrentVersion);
-      const versionIdBeforeDelete = beforeDelete.versionId;
-
-      // Delete the blob (without version ID = delete current)
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
-
-      // Current version should no longer exist
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          undefined,
-          undefined
-        );
-        assert.fail("Should have thrown error for deleted current blob");
-      } catch (error) {
-        // Expected - no current version after delete
-      }
-
-      // Previous version should still be accessible by version ID
-      assert.ok(!isNullOrWhitespace(versionIdBeforeDelete));
-      const previousVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        versionIdBeforeDelete
-      );
-      assert.strictEqual(previousVersion.isCurrentVersion, false);
-      assert.strictEqual(previousVersion.versionId, versionIdBeforeDelete);
-    });
-
-    it("should allow creating new current version after deletion @loki", async () => {
-      const name = `blob-${uuid()}`;
-
-      // Create and delete a version
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "content1");
-      await store.createBlob(ctx, v1);
-      const beforeDelete = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      const deletedVersionId = beforeDelete.versionId;
-
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
-
-      // Create new blob with same name
-      ctx.startTime = new Date(Date.now() + 100);
-      const v2 = buildBlockBlob(ACCOUNT, containerName, name, "content2");
-      await store.createBlob(ctx, v2);
-
-      // New blob should be current version
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(current.isCurrentVersion);
-      assert.strictEqual(
-        current.properties.contentLength,
-        v2.properties.contentLength
-      );
-      assert.notStrictEqual(current.versionId, deletedVersionId);
-
-      // Previous version should still exist as non-current
-      assert.ok(!isNullOrWhitespace(deletedVersionId));
-      const previous = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        deletedVersionId
-      );
-      assert.strictEqual(previous.isCurrentVersion, false);
-    });
-
-    it("should allow deleting specific versions by versionId @loki", async () => {
-      const name = `blob-${uuid()}`;
-
-      // Create multiple versions
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "version1");
-      const created1 = await store.createBlob(ctx, v1);
-      const version1Id = created1.versionId;
-
-      ctx.startTime = new Date(Date.now() + 100);
-      const v2 = buildBlockBlob(ACCOUNT, containerName, name, "version2");
-      await store.createBlob(ctx, v2);
-
-      // Delete specific version (v1) by versionId
-      assert.ok(!isNullOrWhitespace(version1Id));
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-        versionId: version1Id
-      });
-
-      // Current version (v2) should still exist
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(current.isCurrentVersion);
-      assert.strictEqual(
-        current.properties.contentLength,
-        v2.properties.contentLength
-      );
-
-      // Deleted version should no longer be accessible
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          undefined,
-          version1Id
-        );
-        assert.fail("Should have thrown error for deleted version");
-      } catch (error) {
-        // Expected behavior
-      }
-    });
-
-    it("should create versions for write operations on existing blobs @loki", async () => {
-      const name = `blob-${uuid()}`;
-
-      // Create initial blob
-      const initialBlob = buildBlockBlob(
-        ACCOUNT,
-        containerName,
-        name,
-        "initial"
-      );
-      await store.createBlob(ctx, initialBlob);
-      const firstVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Update metadata (write operation) should create new version
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { custommeta: "value" }
-      );
-
-      const afterMetadataUpdate = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should have new version ID and be current
-      assert.notStrictEqual(
-        afterMetadataUpdate.versionId,
-        firstVersion.versionId
-      );
-      assert.ok(afterMetadataUpdate.isCurrentVersion);
-
-      // Previous version should still exist as non-current
-      assert.ok(!isNullOrWhitespace(firstVersion.versionId));
-      const previousVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        firstVersion.versionId
-      );
-      assert.strictEqual(previousVersion.isCurrentVersion, false);
-      assert.strictEqual(previousVersion.versionId, firstVersion.versionId);
-    });
-
-    it("should handle immutable versions correctly @loki", async () => {
-      const name = `blob-${uuid()}`;
-
-      // Create version
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      const created = await store.createBlob(ctx, v1);
-      const versionId = created.versionId;
-
-      // Create new current version
-      ctx.startTime = new Date(Date.now() + 100);
-      const v2 = buildBlockBlob(ACCOUNT, containerName, name, "modified");
-      await store.createBlob(ctx, v2);
-
-      // Previous version should remain unchanged when accessed
-      assert.ok(!isNullOrWhitespace(versionId));
-      const version1_read1 = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        versionId
-      );
-
-      // Wait and read again - should be identical
-      const version1_read2 = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        versionId
-      );
-
-      assert.strictEqual(
-        version1_read1.properties.contentLength,
-        version1_read2.properties.contentLength
-      );
-      assert.strictEqual(
-        version1_read1.properties.etag,
-        version1_read2.properties.etag
-      );
-      assert.strictEqual(version1_read1.versionId, version1_read2.versionId);
-      assert.strictEqual(version1_read1.isCurrentVersion, false);
-      assert.strictEqual(version1_read2.isCurrentVersion, false);
-    });
-
-    it("should return correct isCurrentVersion flag for different scenarios @loki", async () => {
-      const name = `blob-${uuid()}`;
-
-      // Create first version
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "v1");
-      await store.createBlob(ctx, v1);
-
-      // Should be current
-      const first = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(first.isCurrentVersion, true);
-
-      // Create second version
-      ctx.startTime = new Date(Date.now() + 100);
-      const v2 = buildBlockBlob(ACCOUNT, containerName, name, "v2");
-      await store.createBlob(ctx, v2);
-
-      // Second should be current, first should not be
-      const second = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(second.isCurrentVersion, true);
-
-      assert.ok(!isNullOrWhitespace(first.versionId));
-      const firstAgain = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        first.versionId
-      );
-      assert.strictEqual(firstAgain.isCurrentVersion, false);
-    });
-
-    // ================== SNAPSHOT TESTS ==================
-    it("should create a new version when taking a snapshot while versioning enabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      const beforeSnapshot = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Take snapshot should create new version according to Azure docs
-      ctx.startTime = new Date(Date.now() + 100);
-      const snapshotResponse = await store.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-
-      assert.ok(snapshotResponse.snapshot);
-      assert.ok(!isNullOrWhitespace(snapshotResponse.versionId));
-
-      // Current version should have changed after snapshot
-      const afterSnapshot = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      assert.notStrictEqual(afterSnapshot.versionId, beforeSnapshot.versionId);
-      assert.ok(afterSnapshot.isCurrentVersion);
-    });
-
-    // ================== HTTP HEADERS TESTS ==================
-    it("should NOT create new version when setting HTTP headers with versioning enabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      const beforeHeaders = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.setBlobHTTPHeaders(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { blobContentType: "text/plain" }
-      );
-
-      const afterHeaders = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should NOT create new version - HTTP headers are metadata updates only
-      assert.strictEqual(afterHeaders.versionId, beforeHeaders.versionId);
-      assert.strictEqual(afterHeaders.properties.contentType, "text/plain");
-      assert.ok(afterHeaders.isCurrentVersion);
-
-      // Should be the same version with updated headers
-      assert.strictEqual(afterHeaders.versionId, beforeHeaders.versionId);
-    });
-
-    // ================== BLOB TAGS TESTS ==================
-    it("should NOT create new version when setting blob tags with versioning enabled @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      const beforeTags = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.setBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined,
-        { blobTagSet: [{ key: "key1", value: "value1" }] }
-      );
-
-      const afterTags = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should NOT create new version - tags are metadata updates only
-      assert.strictEqual(afterTags.versionId, beforeTags.versionId);
-      assert.ok(afterTags.isCurrentVersion);
-
-      // Verify tags are set on current version (same version)
-      const tags = await store.getBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-      assert.deepStrictEqual(tags, {
-        blobTagSet: [{ key: "key1", value: "value1" }]
-      });
-    });
-
-    it("should access tags from specific versions @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      // Set tags on first version - this should NOT create a new version
-      await store.setBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined,
-        { blobTagSet: [{ key: "version", value: "1" }] }
-      );
-
-      const firstVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Create second version using content change (Put Blob operation)
-      ctx.startTime = new Date(Date.now() + 100);
-      const blob2 = buildBlockBlob(ACCOUNT, containerName, name, "content2");
-      await store.createBlob(ctx, blob2);
-
-      // Set different tags on second version
-      await store.setBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined,
-        { blobTagSet: [{ key: "version", value: "2" }] }
-      );
-
-      // Verify each version has its own tags
-      const firstVersionTags = await store.getBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        firstVersion.versionId,
-        undefined
-      );
-
-      const currentTags = await store.getBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-
-      assert.deepStrictEqual(firstVersionTags, {
-        blobTagSet: [{ key: "version", value: "1" }]
-      });
-      assert.deepStrictEqual(currentTags, {
-        blobTagSet: [{ key: "version", value: "2" }]
-      });
-    });
-
-    // ================== TIER MANAGEMENT TESTS ==================
-    it("should set tier on specific blob versions independently @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      blob.properties.accessTier = Models.AccessTier.Hot;
-      await store.createBlob(ctx, blob);
-
-      const firstVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Create second version
-      ctx.startTime = new Date(Date.now() + 100);
-      const blob2 = buildBlockBlob(ACCOUNT, containerName, name, "content2");
-      blob2.properties.accessTier = Models.AccessTier.Hot;
-      await store.createBlob(ctx, blob2);
-
-      // Set tier on current version (without versionId)
-      await store.setTier(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        "",
-        Models.AccessTier.Cool,
-        undefined
-      );
-
-      // Current version should have Cool tier
-      const currentAfterTier = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(
-        currentAfterTier.properties.accessTier,
-        Models.AccessTier.Cool
-      );
-
-      // Previous version should still have Hot tier
-      assert.ok(!isNullOrWhitespace(firstVersion.versionId));
-      const previousAfterTier = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        firstVersion.versionId
-      );
-      assert.strictEqual(
-        previousAfterTier.properties.accessTier,
-        Models.AccessTier.Hot
-      );
-
-      // Now set tier on specific version (first version) by versionId
-      await store.setTier(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        firstVersion.versionId,
-        Models.AccessTier.Archive,
-        undefined
-      );
-
-      // First version should now have Archive tier
-      const firstVersionAfterArchive = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        firstVersion.versionId
-      );
-      assert.strictEqual(
-        firstVersionAfterArchive.properties.accessTier,
-        Models.AccessTier.Archive
-      );
-
-      // Current version should still have Cool tier (unchanged)
-      const currentStillCool = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(
-        currentStillCool.properties.accessTier,
-        Models.AccessTier.Cool
-      );
-    });
-
-    // ================== BLOB EXISTENCE AND PROPERTIES TESTS ==================
-    it("should check blob existence for specific versions @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      const firstVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Create second version
-      ctx.startTime = new Date(Date.now() + 100);
-      const blob2 = buildBlockBlob(ACCOUNT, containerName, name, "content2");
-      await store.createBlob(ctx, blob2);
-
-      // Check existence of current version
-      await store.checkBlobExist(ctx, ACCOUNT, containerName, name);
-
-      // Check existence of specific version
-      assert.ok(!isNullOrWhitespace(firstVersion.versionId));
-      await store.checkBlobExist(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        "",
-        firstVersion.versionId
-      );
-
-      // Should throw for non-existent version
-      try {
-        await store.checkBlobExist(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          "",
-          "2099-01-01T00:00:00.0000000Z"
-        );
-        assert.fail("Should have thrown for non-existent version");
-      } catch (error) {
-        // Expected
-      }
-    });
-
-    it("should get properties for specific blob versions @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
-
-      // Set metadata to create version
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { version: "1" }
-      );
-
-      const firstVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Set different metadata to create second version
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { version: "2" }
-      );
-
-      // Get properties of current version
-      const currentProps = await store.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-
-      assert.deepStrictEqual(currentProps.metadata, { version: "2" });
-
-      // Get properties of previous version
-      assert.ok(!isNullOrWhitespace(firstVersion.versionId));
-      const prevProps = await store.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        firstVersion.versionId,
-        undefined
-      );
-
-      assert.deepStrictEqual(prevProps.metadata, { version: "1" });
-    });
-
-    // ================== APPEND BLOB OPERATIONS TESTS ==================
-    it("should not create versions for Append Block operations @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const appendBlob = buildAppendBlob(ACCOUNT, containerName, name);
-      await store.createBlob(ctx, appendBlob);
-
-      const afterCreate = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Append block should not create version
-      const block = {
-        accountName: ACCOUNT,
-        containerName,
-        blobName: name,
-        name: "append1",
-        size: 10,
-        persistency: { id: uuid(), offset: 0, count: 10 }
-      } as any;
-
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.appendBlock(ctx, block);
-
-      const afterAppend = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should be same version, just updated properties
-      assert.strictEqual(afterAppend.versionId, afterCreate.versionId);
-      assert.ok(afterAppend.isCurrentVersion);
-      assert.strictEqual(afterAppend.properties.contentLength, 10);
-    });
-
-    it("should create versions for Put Blob operations on append blobs @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const appendBlob1 = buildAppendBlob(ACCOUNT, containerName, name);
-      await store.createBlob(ctx, appendBlob1);
-
-      const firstVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Put Blob (replace) should create new version
-      ctx.startTime = new Date(Date.now() + 100);
-      const appendBlob2 = buildAppendBlob(ACCOUNT, containerName, name);
-      appendBlob2.properties.contentLength = 20;
-      await store.createBlob(ctx, appendBlob2);
-
-      const secondVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      assert.notStrictEqual(secondVersion.versionId, firstVersion.versionId);
-      assert.ok(secondVersion.isCurrentVersion);
-      assert.strictEqual(secondVersion.properties.contentLength, 20);
-    });
-
-    // ================== PAGE BLOB OPERATIONS TESTS ==================
-    it("should not create versions for Put Page operations @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const pageBlob = buildPageBlob(ACCOUNT, containerName, name, 512);
-      await store.createBlob(ctx, pageBlob);
-
-      const afterCreate = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Put Page should not create version according to Azure docs
-      const persistency = { id: uuid(), offset: 0, count: 512 };
-      ctx.startTime = new Date(Date.now() + 100);
-      await store.uploadPages(ctx, pageBlob, 0, 511, persistency);
-
-      const afterUpload = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Should be same version
-      assert.strictEqual(afterUpload.versionId, afterCreate.versionId);
-      assert.ok(afterUpload.isCurrentVersion);
-    });
-
-    it("should create versions for Put Blob operations on page blobs @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const pageBlob1 = buildPageBlob(ACCOUNT, containerName, name, 512);
-      await store.createBlob(ctx, pageBlob1);
-
-      const firstVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Put Blob (replace) should create new version
-      ctx.startTime = new Date(Date.now() + 100);
-      const pageBlob2 = buildPageBlob(ACCOUNT, containerName, name, 1024);
-      await store.createBlob(ctx, pageBlob2);
-
-      const secondVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      assert.notStrictEqual(secondVersion.versionId, firstVersion.versionId);
-      assert.ok(secondVersion.isCurrentVersion);
-      assert.strictEqual(secondVersion.properties.contentLength, 1024);
-    });
-
-    // ================== VERSION MODE TRANSITION TESTS ==================
-    it("should handle setBlobMetadata correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create base blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await disabledStore.createBlob(ctx, baseBlob);
-
-      // Set metadata on base blob (should not create version)
-      ctx.startTime = new Date(Date.now() + 100);
-      await disabledStore.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { basemeta: "value1" }
-      );
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      assert.deepStrictEqual(baseFetched.metadata, { basemeta: "value1" });
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Set metadata should create new version and promote previous
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { versionedmeta: "value2" }
-      );
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(!isNullOrWhitespace(current.versionId));
-      assert.ok(current.isCurrentVersion);
-      assert.deepStrictEqual(current.metadata, { versionedmeta: "value2" });
-
-      // Previous version should be accessible with original metadata
-      const originalLastModifiedIso =
-        baseFetched.properties.lastModified.toISOString();
-      const previous = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        originalLastModifiedIso
-      );
-      assert.strictEqual(previous.isCurrentVersion, false);
-      assert.deepStrictEqual(previous.metadata, { basemeta: "value1" });
-    });
-
-    it("should handle setBlobHTTPHeaders correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create base blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await disabledStore.createBlob(ctx, baseBlob);
-
-      // Set HTTP headers on base blob (should not create version)
-      ctx.startTime = new Date(Date.now() + 100);
-      await disabledStore.setBlobHTTPHeaders(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { blobContentType: "text/plain" }
-      );
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      assert.strictEqual(baseFetched.properties.contentType, "text/plain");
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Set headers should NOT create new version (metadata operation)
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setBlobHTTPHeaders(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { blobContentType: "application/json" }
-      );
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should be promoted version but same version (no new version for headers)
-      assert.strictEqual(current.versionId, "");
-      assert.ok(!current.isCurrentVersion);
-      assert.strictEqual(current.properties.contentType, "application/json");
-    });
-
-    it("should handle setBlobTag/getBlobTag correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create base blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await disabledStore.createBlob(ctx, baseBlob);
-
-      // Set tags on base blob (should not create version)
-      ctx.startTime = new Date(Date.now() + 100);
-      await disabledStore.setBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined,
-        { blobTagSet: [{ key: "env", value: "test" }] }
-      );
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      const baseTags = await disabledStore.getBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-      assert.deepStrictEqual(baseTags, {
-        blobTagSet: [{ key: "env", value: "test" }]
-      });
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Set tags should NOT create new version (metadata operation)
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined,
-        { blobTagSet: [{ key: "env", value: "prod" }] }
-      );
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should be promoted version but same version (no new version for tags)
-      assert.strictEqual(current.versionId, "");
-      assert.ok(!current.isCurrentVersion);
-
-      const currentTags = await store.getBlobTag(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-      assert.deepStrictEqual(currentTags, {
-        blobTagSet: [{ key: "env", value: "prod" }]
-      });
-    });
-
-    it("should handle setTier correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create base blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      baseBlob.properties.accessTier = Models.AccessTier.Hot;
-      await disabledStore.createBlob(ctx, baseBlob);
-
-      // Set tier on base blob (should not create version)
-      ctx.startTime = new Date(Date.now() + 100);
-      await disabledStore.setTier(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        Models.AccessTier.Cool,
-        undefined
-      );
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      assert.strictEqual(
-        baseFetched.properties.accessTier,
-        Models.AccessTier.Cool
-      );
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Set tier should work on promoted version
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setTier(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        Models.AccessTier.Archive,
-        undefined
-      );
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should be promoted version
-      assert.strictEqual(current.versionId, "");
-      assert.ok(!current.isCurrentVersion);
-      assert.strictEqual(
-        current.properties.accessTier,
-        Models.AccessTier.Archive
-      );
-    });
-
-    it("should handle checkBlobExist correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create base blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await disabledStore.createBlob(ctx, baseBlob);
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      const originalLastModifiedIso =
-        baseFetched.properties.lastModified.toISOString();
-
-      // Check existence should work
-      await disabledStore.checkBlobExist(ctx, ACCOUNT, containerName, name);
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Check existence should work for promoted base blob
-      await store.checkBlobExist(ctx, ACCOUNT, containerName, name);
-
-      // No writes yet, so versionId should be empty
-      await store.checkBlobExist(ctx, ACCOUNT, containerName, name, "", "");
-
-      // Create new version to ensure previous works
-      ctx.startTime = new Date(Date.now() + 200);
-      const newBlob = buildBlockBlob(ACCOUNT, containerName, name, "new");
-      await store.createBlob(ctx, newBlob);
-
-      // Both current and previous should exist
-      await store.checkBlobExist(ctx, ACCOUNT, containerName, name);
-      await store.checkBlobExist(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        "",
-        originalLastModifiedIso
-      );
-    });
-
-    it("should handle getBlobProperties correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create base blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await disabledStore.createBlob(ctx, baseBlob);
-
-      // Set metadata
-      ctx.startTime = new Date(Date.now() + 100);
-      await disabledStore.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { env: "test" }
-      );
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      const originalLastModifiedIso =
-        baseFetched.properties.lastModified.toISOString();
-
-      // Get properties should work
-      const baseProps = await disabledStore.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-      assert.deepStrictEqual(baseProps.metadata, { env: "test" });
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Get properties should work for promoted base blob
-      const promotedProps = await store.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-      assert.deepStrictEqual(promotedProps.metadata, { env: "test" });
-
-      // Create new version
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.setBlobMetadata(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        { env: "prod" }
-      );
-
-      // Get current properties
-      const currentProps = await store.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined,
-        undefined
-      );
-      assert.deepStrictEqual(currentProps.metadata, { env: "prod" });
-
-      // Get previous version properties
-      const prevProps = await store.getBlobProperties(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        originalLastModifiedIso,
-        undefined
-      );
-      assert.deepStrictEqual(prevProps.metadata, { env: "test" });
-    });
-
-    it("should handle createSnapshot correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create base blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await disabledStore.createBlob(ctx, baseBlob);
-
-      // Create snapshot (should not create version)
-      ctx.startTime = new Date(Date.now() + 100);
-      const snapshotResponse1 = await disabledStore.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-      assert.ok(snapshotResponse1.snapshot);
-      assert.strictEqual(snapshotResponse1.versionId, "");
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      const originalLastModifiedIso =
-        baseFetched.properties.lastModified.toISOString();
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Create snapshot should create new version and promote previous
-      ctx.startTime = new Date(Date.now() + 200);
-      const snapshotResponse2 = await store.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-      assert.ok(snapshotResponse2.snapshot);
-      assert.ok(!isNullOrWhitespace(snapshotResponse2.versionId));
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.ok(!isNullOrWhitespace(current.versionId));
-      assert.ok(current.isCurrentVersion);
-
-      // Previous version should be accessible
-      const previous = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        originalLastModifiedIso
-      );
-      assert.strictEqual(previous.isCurrentVersion, false);
-      assert.strictEqual(previous.versionId, originalLastModifiedIso);
-    });
-
-    it("should handle appendBlock correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create append blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseAppendBlob = buildAppendBlob(ACCOUNT, containerName, name);
-      await disabledStore.createBlob(ctx, baseAppendBlob);
-
-      // Append block (should not create version)
-      const block1 = {
-        accountName: ACCOUNT,
-        containerName,
-        blobName: name,
-        name: "append1",
-        size: 10,
-        persistency: { id: uuid(), offset: 0, count: 10 }
-      } as any;
-
-      ctx.startTime = new Date(Date.now() + 100);
-      await disabledStore.appendBlock(ctx, block1);
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      assert.strictEqual(baseFetched.properties.contentLength, 10);
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Append block should NOT create new version (per Azure spec)
-      const block2 = {
-        accountName: ACCOUNT,
-        containerName,
-        blobName: name,
-        name: "append2",
-        size: 15,
-        persistency: { id: uuid(), offset: 10, count: 15 }
-      } as any;
-
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.appendBlock(ctx, block2);
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should not be promoted version (no new version for page upload).
-      // It does not count as a versioning "write."
-      assert.strictEqual(current.versionId, "");
-      assert.ok(!current.isCurrentVersion);
-      assert.strictEqual(current.properties.contentLength, 25);
-    });
-
-    it("should handle uploadPages correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create page blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const basePageBlob = buildPageBlob(ACCOUNT, containerName, name, 512);
-      await disabledStore.createBlob(ctx, basePageBlob);
-
-      // Upload pages (should not create version)
-      const persistency1 = { id: uuid(), offset: 0, count: 512 };
-      ctx.startTime = new Date(Date.now() + 100);
-      await disabledStore.uploadPages(ctx, basePageBlob, 0, 511, persistency1);
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Upload pages should NOT create new version (per Azure spec)
-      const persistency2 = { id: uuid(), offset: 0, count: 512 };
-      ctx.startTime = new Date(Date.now() + 200);
-      await store.uploadPages(ctx, basePageBlob, 0, 511, persistency2);
-
-      const current = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      // Should not be promoted version (no new version for page upload).
-      // It does not count as a versioning "write."
-      assert.strictEqual(current.versionId, "");
-      assert.ok(!current.isCurrentVersion);
-    });
-
-    it("should handle deleteBlob correctly across versioning mode transitions @loki", async () => {
-      await store.close();
-      await store.clean();
-
-      const name = `blob-${uuid()}`;
-
-      // 1. Create store with versioning DISABLED and create base blob
-      let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
-      const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
-      await disabledStore.createBlob(ctx, baseBlob);
-
-      const baseFetched = await disabledStore.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-      assert.strictEqual(baseFetched.versionId, "");
-      const originalLastModifiedIso =
-        baseFetched.properties.lastModified.toISOString();
-      await disabledStore.close();
-
-      // 2. Re-open with versioning ENABLED
-      store = new LokiBlobMetadataStore(DB_FILE, false, true);
-      await store.init();
-
-      // Create new version first so we have something to delete
-      ctx.startTime = new Date(Date.now() + 200);
-      const newBlob = buildBlockBlob(ACCOUNT, containerName, name, "new");
-      await store.createBlob(ctx, newBlob);
-
-      const currentBeforeDelete = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        undefined
-      );
-
-      // Delete current blob should make it non-current
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
-
-      // Current version should no longer exist
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          undefined,
-          undefined
-        );
-        assert.fail("Should have thrown error for deleted current blob");
-      } catch (error) {
-        // Expected
-      }
-
-      // Previous versions should still exist as non-current
-      const originalVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        originalLastModifiedIso
-      );
-      assert.strictEqual(originalVersion.isCurrentVersion, false);
-
-      const deletedVersion = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name,
-        undefined,
-        currentBeforeDelete.versionId
-      );
-      assert.strictEqual(deletedVersion.isCurrentVersion, false);
-
-      // Should be able to delete specific version by versionId
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-        versionId: originalLastModifiedIso
-      });
-
-      // That specific version should no longer exist
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          undefined,
-          originalLastModifiedIso
-        );
-        assert.fail("Should have thrown error for deleted specific version");
-      } catch (error) {
-        // Expected
-      }
+    const afterMetadataUpdate = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Should have new version ID and be current
+    assert.notStrictEqual(
+      afterMetadataUpdate.versionId,
+      firstVersion.versionId
+    );
+    assert.ok(afterMetadataUpdate.isCurrentVersion);
+
+    // Previous version should still exist as non-current
+    assert.ok(!isNullOrWhitespace(firstVersion.versionId));
+    const previousVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      firstVersion.versionId
+    );
+    assert.strictEqual(previousVersion.isCurrentVersion, false);
+    assert.strictEqual(previousVersion.versionId, firstVersion.versionId);
+  });
+
+  it("should handle immutable versions correctly @loki", async () => {
+    const name = `blob-${uuid()}`;
+
+    // Create version
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    const created = await store.createBlob(ctx, v1);
+    const versionId = created.versionId;
+
+    // Create new current version
+    ctx.startTime = new Date(Date.now() + 100);
+    const v2 = buildBlockBlob(ACCOUNT, containerName, name, "modified");
+    await store.createBlob(ctx, v2);
+
+    // Previous version should remain unchanged when accessed
+    assert.ok(!isNullOrWhitespace(versionId));
+    const version1_read1 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      versionId
+    );
+
+    // Wait and read again - should be identical
+    const version1_read2 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      versionId
+    );
+
+    assert.strictEqual(
+      version1_read1.properties.contentLength,
+      version1_read2.properties.contentLength
+    );
+    assert.strictEqual(
+      version1_read1.properties.etag,
+      version1_read2.properties.etag
+    );
+    assert.strictEqual(version1_read1.versionId, version1_read2.versionId);
+    assert.strictEqual(version1_read1.isCurrentVersion, false);
+    assert.strictEqual(version1_read2.isCurrentVersion, false);
+  });
+
+  it("should return correct isCurrentVersion flag for different scenarios @loki", async () => {
+    const name = `blob-${uuid()}`;
+
+    // Create first version
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "v1");
+    await store.createBlob(ctx, v1);
+
+    // Should be current
+    const first = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(first.isCurrentVersion, true);
+
+    // Create second version
+    ctx.startTime = new Date(Date.now() + 100);
+    const v2 = buildBlockBlob(ACCOUNT, containerName, name, "v2");
+    await store.createBlob(ctx, v2);
+
+    // Second should be current, first should not be
+    const second = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(second.isCurrentVersion, true);
+
+    assert.ok(!isNullOrWhitespace(first.versionId));
+    const firstAgain = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      first.versionId
+    );
+    assert.strictEqual(firstAgain.isCurrentVersion, false);
+  });
+
+  // ================== SNAPSHOT TESTS ==================
+  it("should create a new version when taking a snapshot while versioning enabled @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
+
+    const beforeSnapshot = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Take snapshot should create new version according to Azure docs
+    ctx.startTime = new Date(Date.now() + 100);
+    const snapshotResponse = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+
+    assert.ok(snapshotResponse.snapshot);
+    assert.ok(!isNullOrWhitespace(snapshotResponse.versionId));
+
+    // Current version should have changed after snapshot
+    const afterSnapshot = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    assert.notStrictEqual(afterSnapshot.versionId, beforeSnapshot.versionId);
+    assert.ok(afterSnapshot.isCurrentVersion);
+  });
+
+  // ================== HTTP HEADERS TESTS ==================
+  it("should NOT create new version when setting HTTP headers with versioning enabled @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
+
+    const beforeHeaders = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    ctx.startTime = new Date(Date.now() + 100);
+    await store.setBlobHTTPHeaders(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { blobContentType: "text/plain" }
+    );
+
+    const afterHeaders = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Should NOT create new version - HTTP headers are metadata updates only
+    assert.strictEqual(afterHeaders.versionId, beforeHeaders.versionId);
+    assert.strictEqual(afterHeaders.properties.contentType, "text/plain");
+    assert.ok(afterHeaders.isCurrentVersion);
+
+    // Should be the same version with updated headers
+    assert.strictEqual(afterHeaders.versionId, beforeHeaders.versionId);
+  });
+
+  // ================== BLOB TAGS TESTS ==================
+  it("should NOT create new version when setting blob tags with versioning enabled @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
+
+    const beforeTags = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    ctx.startTime = new Date(Date.now() + 100);
+    await store.setBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined,
+      { blobTagSet: [{ key: "key1", value: "value1" }] }
+    );
+
+    const afterTags = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Should NOT create new version - tags are metadata updates only
+    assert.strictEqual(afterTags.versionId, beforeTags.versionId);
+    assert.ok(afterTags.isCurrentVersion);
+
+    // Verify tags are set on current version (same version)
+    const tags = await store.getBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined
+    );
+    assert.deepStrictEqual(tags, {
+      blobTagSet: [{ key: "key1", value: "value1" }]
     });
   });
 
-  describe("deleteBlob comprehensive code path coverage @loki", () => {
-    let store: LokiBlobMetadataStore;
-    let disabledStore: LokiBlobMetadataStore;
-    let ctx: Context;
-    const containerName = "test-container";
+  it("should access tags from specific versions @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
 
-    beforeEach(async () => {
-      ctx = createContext();
-      // Versioning enabled
-      store = new LokiBlobMetadataStore("__test_db_blob__.json", false, true);
-      await store.init();
-      await store.createContainer(ctx, buildContainer(ACCOUNT, containerName));
+    // Set tags on first version - this should NOT create a new version
+    await store.setBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined,
+      { blobTagSet: [{ key: "version", value: "1" }] }
+    );
 
-      // Versioning disabled
-      disabledStore = new LokiBlobMetadataStore(
-        "__test_db_blob_disabled__.json",
-        false,
-        false
-      );
-      await disabledStore.init();
-      await disabledStore.createContainer(
-        ctx,
-        buildContainer(ACCOUNT, containerName)
-      );
+    const firstVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Create second version using content change (Put Blob operation)
+    ctx.startTime = new Date(Date.now() + 100);
+    const blob2 = buildBlockBlob(ACCOUNT, containerName, name, "content2");
+    await store.createBlob(ctx, blob2);
+
+    // Set different tags on second version
+    await store.setBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined,
+      { blobTagSet: [{ key: "version", value: "2" }] }
+    );
+
+    // Verify each version has its own tags
+    const firstVersionTags = await store.getBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      firstVersion.versionId,
+      undefined
+    );
+
+    const currentTags = await store.getBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined
+    );
+
+    assert.deepStrictEqual(firstVersionTags, {
+      blobTagSet: [{ key: "version", value: "1" }]
     });
-
-    afterEach(async () => {
-      await store.close();
-      await store.clean();
-      await disabledStore.close();
-      await disabledStore.clean();
+    assert.deepStrictEqual(currentTags, {
+      blobTagSet: [{ key: "version", value: "2" }]
     });
+  });
 
-    it("should throw error when versionId is provided with snapshot option @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      const created = await store.createBlob(ctx, blob);
+  // ================== TIER MANAGEMENT TESTS ==================
+  it("should set tier on specific blob versions independently @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    blob.properties.accessTier = Models.AccessTier.Hot;
+    await store.createBlob(ctx, blob);
 
-      // Create a snapshot
-      const snapshot = await store.createSnapshot(
+    const firstVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Create second version
+    ctx.startTime = new Date(Date.now() + 100);
+    const blob2 = buildBlockBlob(ACCOUNT, containerName, name, "content2");
+    blob2.properties.accessTier = Models.AccessTier.Hot;
+    await store.createBlob(ctx, blob2);
+
+    // Set tier on current version (without versionId)
+    await store.setTier(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      "",
+      Models.AccessTier.Cool,
+      undefined
+    );
+
+    // Current version should have Cool tier
+    const currentAfterTier = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(
+      currentAfterTier.properties.accessTier,
+      Models.AccessTier.Cool
+    );
+
+    // Previous version should still have Hot tier
+    assert.ok(!isNullOrWhitespace(firstVersion.versionId));
+    const previousAfterTier = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      firstVersion.versionId
+    );
+    assert.strictEqual(
+      previousAfterTier.properties.accessTier,
+      Models.AccessTier.Hot
+    );
+
+    // Now set tier on specific version (first version) by versionId
+    await store.setTier(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      firstVersion.versionId,
+      Models.AccessTier.Archive,
+      undefined
+    );
+
+    // First version should now have Archive tier
+    const firstVersionAfterArchive = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      firstVersion.versionId
+    );
+    assert.strictEqual(
+      firstVersionAfterArchive.properties.accessTier,
+      Models.AccessTier.Archive
+    );
+
+    // Current version should still have Cool tier (unchanged)
+    const currentStillCool = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(
+      currentStillCool.properties.accessTier,
+      Models.AccessTier.Cool
+    );
+  });
+
+  // ================== BLOB EXISTENCE AND PROPERTIES TESTS ==================
+  it("should check blob existence for specific versions @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
+
+    const firstVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Create second version
+    ctx.startTime = new Date(Date.now() + 100);
+    const blob2 = buildBlockBlob(ACCOUNT, containerName, name, "content2");
+    await store.createBlob(ctx, blob2);
+
+    // Check existence of current version
+    await store.checkBlobExist(ctx, ACCOUNT, containerName, name);
+
+    // Check existence of specific version
+    assert.ok(!isNullOrWhitespace(firstVersion.versionId));
+    await store.checkBlobExist(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      "",
+      firstVersion.versionId
+    );
+
+    // Should throw for non-existent version
+    try {
+      await store.checkBlobExist(
         ctx,
         ACCOUNT,
         containerName,
-        name
+        name,
+        "",
+        "2099-01-01T00:00:00.0000000Z"
       );
+      assert.fail("Should have thrown for non-existent version");
+    } catch (error) {
+      // Expected
+    }
+  });
 
-      try {
-        await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-          snapshot: snapshot.snapshot,
-          versionId: created.versionId
-        });
-        assert.fail(
-          "Should have thrown error when versionId provided with snapshot"
-        );
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 400);
-        assert.ok(
-          error.message.includes(
-            "When deleting a blob version, you cannot specify a snapshot"
-          )
-        );
-      }
+  it("should get properties for specific blob versions @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
+
+    // Set metadata to create version
+    ctx.startTime = new Date(Date.now() + 100);
+    await store.setBlobMetadata(ctx, ACCOUNT, containerName, name, undefined, {
+      version: "1"
     });
 
-    it("should throw error when versionId is provided with deleteSnapshots option @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      const created = await store.createBlob(ctx, blob);
+    const firstVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
 
-      try {
-        await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-          deleteSnapshots: Models.DeleteSnapshotsOptionType.Include,
-          versionId: created.versionId
-        });
-        assert.fail(
-          "Should have thrown error when versionId provided with deleteSnapshots"
-        );
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 400);
-        assert.ok(
-          error.message.includes(
-            "When deleting a blob version, you cannot specify a snapshot"
-          )
-        );
-      }
+    // Set different metadata to create second version
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setBlobMetadata(ctx, ACCOUNT, containerName, name, undefined, {
+      version: "2"
     });
 
-    it("should throw BlobNotFound when deleting non-existent blob @loki", async () => {
-      try {
-        await store.deleteBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          "non-existent-blob",
-          {}
-        );
-        assert.fail("Should have thrown BlobNotFound error");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
+    // Get properties of current version
+    const currentProps = await store.getBlobProperties(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined
+    );
+
+    assert.deepStrictEqual(currentProps.metadata, { version: "2" });
+
+    // Get properties of previous version
+    assert.ok(!isNullOrWhitespace(firstVersion.versionId));
+    const prevProps = await store.getBlobProperties(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      firstVersion.versionId,
+      undefined
+    );
+
+    assert.deepStrictEqual(prevProps.metadata, { version: "1" });
+  });
+
+  // ================== APPEND BLOB OPERATIONS TESTS ==================
+  it("should not create versions for Append Block operations @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const appendBlob = buildAppendBlob(ACCOUNT, containerName, name);
+    await store.createBlob(ctx, appendBlob);
+
+    const afterCreate = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Append block should not create version
+    const block = {
+      accountName: ACCOUNT,
+      containerName,
+      blobName: name,
+      name: "append1",
+      size: 10,
+      persistency: { id: uuid(), offset: 0, count: 10 }
+    } as any;
+
+    ctx.startTime = new Date(Date.now() + 100);
+    await store.appendBlock(ctx, block);
+
+    const afterAppend = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Should be same version, just updated properties
+    assert.strictEqual(afterAppend.versionId, afterCreate.versionId);
+    assert.ok(afterAppend.isCurrentVersion);
+    assert.strictEqual(afterAppend.properties.contentLength, 10);
+  });
+
+  it("should create versions for Put Blob operations on append blobs @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const appendBlob1 = buildAppendBlob(ACCOUNT, containerName, name);
+    await store.createBlob(ctx, appendBlob1);
+
+    const firstVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Put Blob (replace) should create new version
+    ctx.startTime = new Date(Date.now() + 100);
+    const appendBlob2 = buildAppendBlob(ACCOUNT, containerName, name);
+    appendBlob2.properties.contentLength = 20;
+    await store.createBlob(ctx, appendBlob2);
+
+    const secondVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    assert.notStrictEqual(secondVersion.versionId, firstVersion.versionId);
+    assert.ok(secondVersion.isCurrentVersion);
+    assert.strictEqual(secondVersion.properties.contentLength, 20);
+  });
+
+  // ================== PAGE BLOB OPERATIONS TESTS ==================
+  it("should not create versions for Put Page operations @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const pageBlob = buildPageBlob(ACCOUNT, containerName, name, 512);
+    await store.createBlob(ctx, pageBlob);
+
+    const afterCreate = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Put Page should not create version according to Azure docs
+    const persistency = { id: uuid(), offset: 0, count: 512 };
+    ctx.startTime = new Date(Date.now() + 100);
+    await store.uploadPages(ctx, pageBlob, 0, 511, persistency);
+
+    const afterUpload = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Should be same version
+    assert.strictEqual(afterUpload.versionId, afterCreate.versionId);
+    assert.ok(afterUpload.isCurrentVersion);
+  });
+
+  it("should create versions for Put Blob operations on page blobs @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const pageBlob1 = buildPageBlob(ACCOUNT, containerName, name, 512);
+    await store.createBlob(ctx, pageBlob1);
+
+    const firstVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Put Blob (replace) should create new version
+    ctx.startTime = new Date(Date.now() + 100);
+    const pageBlob2 = buildPageBlob(ACCOUNT, containerName, name, 1024);
+    await store.createBlob(ctx, pageBlob2);
+
+    const secondVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    assert.notStrictEqual(secondVersion.versionId, firstVersion.versionId);
+    assert.ok(secondVersion.isCurrentVersion);
+    assert.strictEqual(secondVersion.properties.contentLength, 1024);
+  });
+
+  // ================== VERSION MODE TRANSITION TESTS ==================
+  it("should handle setBlobMetadata correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create base blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await disabledStore.createBlob(ctx, baseBlob);
+
+    // Set metadata on base blob (should not create version)
+    ctx.startTime = new Date(Date.now() + 100);
+    await disabledStore.setBlobMetadata(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { basemeta: "value1" }
+    );
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    assert.deepStrictEqual(baseFetched.metadata, { basemeta: "value1" });
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Set metadata should create new version and promote previous
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setBlobMetadata(ctx, ACCOUNT, containerName, name, undefined, {
+      versionedmeta: "value2"
     });
 
-    it("should throw error when trying to use deleteSnapshots against a snapshot @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(!isNullOrWhitespace(current.versionId));
+    assert.ok(current.isCurrentVersion);
+    assert.deepStrictEqual(current.metadata, { versionedmeta: "value2" });
 
-      // Create a snapshot
-      const snapshot = await store.createSnapshot(
+    // Previous version should be accessible with original metadata
+    const originalLastModifiedIso =
+      baseFetched.properties.lastModified.toISOString();
+    const previous = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      originalLastModifiedIso
+    );
+    assert.strictEqual(previous.isCurrentVersion, false);
+    assert.deepStrictEqual(previous.metadata, { basemeta: "value1" });
+  });
+
+  it("should handle setBlobHTTPHeaders correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create base blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await disabledStore.createBlob(ctx, baseBlob);
+
+    // Set HTTP headers on base blob (should not create version)
+    ctx.startTime = new Date(Date.now() + 100);
+    await disabledStore.setBlobHTTPHeaders(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { blobContentType: "text/plain" }
+    );
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    assert.strictEqual(baseFetched.properties.contentType, "text/plain");
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Set headers should NOT create new version (metadata operation)
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setBlobHTTPHeaders(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { blobContentType: "application/json" }
+    );
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should be promoted version but same version (no new version for headers)
+    assert.strictEqual(current.versionId, "");
+    assert.ok(!current.isCurrentVersion);
+    assert.strictEqual(current.properties.contentType, "application/json");
+  });
+
+  it("should handle setBlobTag/getBlobTag correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create base blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await disabledStore.createBlob(ctx, baseBlob);
+
+    // Set tags on base blob (should not create version)
+    ctx.startTime = new Date(Date.now() + 100);
+    await disabledStore.setBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined,
+      { blobTagSet: [{ key: "env", value: "test" }] }
+    );
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    const baseTags = await disabledStore.getBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined
+    );
+    assert.deepStrictEqual(baseTags, {
+      blobTagSet: [{ key: "env", value: "test" }]
+    });
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Set tags should NOT create new version (metadata operation)
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined,
+      { blobTagSet: [{ key: "env", value: "prod" }] }
+    );
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should be promoted version but same version (no new version for tags)
+    assert.strictEqual(current.versionId, "");
+    assert.ok(!current.isCurrentVersion);
+
+    const currentTags = await store.getBlobTag(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined
+    );
+    assert.deepStrictEqual(currentTags, {
+      blobTagSet: [{ key: "env", value: "prod" }]
+    });
+  });
+
+  it("should handle setTier correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create base blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    baseBlob.properties.accessTier = Models.AccessTier.Hot;
+    await disabledStore.createBlob(ctx, baseBlob);
+
+    // Set tier on base blob (should not create version)
+    ctx.startTime = new Date(Date.now() + 100);
+    await disabledStore.setTier(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      Models.AccessTier.Cool,
+      undefined
+    );
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    assert.strictEqual(
+      baseFetched.properties.accessTier,
+      Models.AccessTier.Cool
+    );
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Set tier should work on promoted version
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setTier(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      Models.AccessTier.Archive,
+      undefined
+    );
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should be promoted version
+    assert.strictEqual(current.versionId, "");
+    assert.ok(!current.isCurrentVersion);
+    assert.strictEqual(
+      current.properties.accessTier,
+      Models.AccessTier.Archive
+    );
+  });
+
+  it("should handle checkBlobExist correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create base blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await disabledStore.createBlob(ctx, baseBlob);
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    const originalLastModifiedIso =
+      baseFetched.properties.lastModified.toISOString();
+
+    // Check existence should work
+    await disabledStore.checkBlobExist(ctx, ACCOUNT, containerName, name);
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Check existence should work for promoted base blob
+    await store.checkBlobExist(ctx, ACCOUNT, containerName, name);
+
+    // No writes yet, so versionId should be empty
+    await store.checkBlobExist(ctx, ACCOUNT, containerName, name, "", "");
+
+    // Create new version to ensure previous works
+    ctx.startTime = new Date(Date.now() + 200);
+    const newBlob = buildBlockBlob(ACCOUNT, containerName, name, "new");
+    await store.createBlob(ctx, newBlob);
+
+    // Both current and previous should exist
+    await store.checkBlobExist(ctx, ACCOUNT, containerName, name);
+    await store.checkBlobExist(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      "",
+      originalLastModifiedIso
+    );
+  });
+
+  it("should handle getBlobProperties correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create base blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await disabledStore.createBlob(ctx, baseBlob);
+
+    // Set metadata
+    ctx.startTime = new Date(Date.now() + 100);
+    await disabledStore.setBlobMetadata(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      { env: "test" }
+    );
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    const originalLastModifiedIso =
+      baseFetched.properties.lastModified.toISOString();
+
+    // Get properties should work
+    const baseProps = await disabledStore.getBlobProperties(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined
+    );
+    assert.deepStrictEqual(baseProps.metadata, { env: "test" });
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Get properties should work for promoted base blob
+    const promotedProps = await store.getBlobProperties(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined
+    );
+    assert.deepStrictEqual(promotedProps.metadata, { env: "test" });
+
+    // Create new version
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.setBlobMetadata(ctx, ACCOUNT, containerName, name, undefined, {
+      env: "prod"
+    });
+
+    // Get current properties
+    const currentProps = await store.getBlobProperties(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined,
+      undefined
+    );
+    assert.deepStrictEqual(currentProps.metadata, { env: "prod" });
+
+    // Get previous version properties
+    const prevProps = await store.getBlobProperties(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      originalLastModifiedIso,
+      undefined
+    );
+    assert.deepStrictEqual(prevProps.metadata, { env: "test" });
+  });
+
+  it("should handle createSnapshot correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create base blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await disabledStore.createBlob(ctx, baseBlob);
+
+    // Create snapshot (should not create version)
+    ctx.startTime = new Date(Date.now() + 100);
+    const snapshotResponse1 = await disabledStore.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    assert.ok(snapshotResponse1.snapshot);
+    assert.strictEqual(snapshotResponse1.versionId, "");
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    const originalLastModifiedIso =
+      baseFetched.properties.lastModified.toISOString();
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Create snapshot should create new version and promote previous
+    ctx.startTime = new Date(Date.now() + 200);
+    const snapshotResponse2 = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    assert.ok(snapshotResponse2.snapshot);
+    assert.ok(!isNullOrWhitespace(snapshotResponse2.versionId));
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.ok(!isNullOrWhitespace(current.versionId));
+    assert.ok(current.isCurrentVersion);
+
+    // Previous version should be accessible
+    const previous = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      originalLastModifiedIso
+    );
+    assert.strictEqual(previous.isCurrentVersion, false);
+    assert.strictEqual(previous.versionId, originalLastModifiedIso);
+  });
+
+  it("should handle appendBlock correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create append blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseAppendBlob = buildAppendBlob(ACCOUNT, containerName, name);
+    await disabledStore.createBlob(ctx, baseAppendBlob);
+
+    // Append block (should not create version)
+    const block1 = {
+      accountName: ACCOUNT,
+      containerName,
+      blobName: name,
+      name: "append1",
+      size: 10,
+      persistency: { id: uuid(), offset: 0, count: 10 }
+    } as any;
+
+    ctx.startTime = new Date(Date.now() + 100);
+    await disabledStore.appendBlock(ctx, block1);
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    assert.strictEqual(baseFetched.properties.contentLength, 10);
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Append block should NOT create new version (per Azure spec)
+    const block2 = {
+      accountName: ACCOUNT,
+      containerName,
+      blobName: name,
+      name: "append2",
+      size: 15,
+      persistency: { id: uuid(), offset: 10, count: 15 }
+    } as any;
+
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.appendBlock(ctx, block2);
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should not be promoted version (no new version for page upload).
+    // It does not count as a versioning "write."
+    assert.strictEqual(current.versionId, "");
+    assert.ok(!current.isCurrentVersion);
+    assert.strictEqual(current.properties.contentLength, 25);
+  });
+
+  it("should handle uploadPages correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create page blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const basePageBlob = buildPageBlob(ACCOUNT, containerName, name, 512);
+    await disabledStore.createBlob(ctx, basePageBlob);
+
+    // Upload pages (should not create version)
+    const persistency1 = { id: uuid(), offset: 0, count: 512 };
+    ctx.startTime = new Date(Date.now() + 100);
+    await disabledStore.uploadPages(ctx, basePageBlob, 0, 511, persistency1);
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Upload pages should NOT create new version (per Azure spec)
+    const persistency2 = { id: uuid(), offset: 0, count: 512 };
+    ctx.startTime = new Date(Date.now() + 200);
+    await store.uploadPages(ctx, basePageBlob, 0, 511, persistency2);
+
+    const current = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    // Should not be promoted version (no new version for page upload).
+    // It does not count as a versioning "write."
+    assert.strictEqual(current.versionId, "");
+    assert.ok(!current.isCurrentVersion);
+  });
+
+  it("should handle deleteBlob correctly across versioning mode transitions @loki", async () => {
+    await store.close();
+    await store.clean();
+
+    const name = `blob-${uuid()}`;
+
+    // 1. Create store with versioning DISABLED and create base blob
+    let disabledStore = new LokiBlobMetadataStore(DB_FILE, false, false);
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+    const baseBlob = buildBlockBlob(ACCOUNT, containerName, name, "base");
+    await disabledStore.createBlob(ctx, baseBlob);
+
+    const baseFetched = await disabledStore.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+    assert.strictEqual(baseFetched.versionId, "");
+    const originalLastModifiedIso =
+      baseFetched.properties.lastModified.toISOString();
+    await disabledStore.close();
+
+    // 2. Re-open with versioning ENABLED
+    store = new LokiBlobMetadataStore(DB_FILE, false, true);
+    await store.init();
+
+    // Create new version first so we have something to delete
+    ctx.startTime = new Date(Date.now() + 200);
+    const newBlob = buildBlockBlob(ACCOUNT, containerName, name, "new");
+    await store.createBlob(ctx, newBlob);
+
+    const currentBeforeDelete = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      undefined
+    );
+
+    // Delete current blob should make it non-current
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
+
+    // Current version should no longer exist
+    try {
+      await store.downloadBlob(
         ctx,
         ACCOUNT,
         containerName,
-        name
+        name,
+        undefined,
+        undefined
       );
+      assert.fail("Should have thrown error for deleted current blob");
+    } catch (error) {
+      // Expected
+    }
 
-      try {
-        await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-          snapshot: snapshot.snapshot,
-          deleteSnapshots: Models.DeleteSnapshotsOptionType.Include
-        });
-        assert.fail(
-          "Should have thrown error when using deleteSnapshots against snapshot"
-        );
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 400);
-        assert.ok(
-          error.message.includes("Invalid operation against a blob snapshot")
-        );
-      }
+    // Previous versions should still exist as non-current
+    const originalVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      originalLastModifiedIso
+    );
+    assert.strictEqual(originalVersion.isCurrentVersion, false);
+
+    const deletedVersion = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      undefined,
+      currentBeforeDelete.versionId
+    );
+    assert.strictEqual(deletedVersion.isCurrentVersion, false);
+
+    // Should be able to delete specific version by versionId
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
+      versionId: originalLastModifiedIso
     });
 
-    it("should delete specific version when versionId is provided @loki", async () => {
-      const name = `blob-${uuid()}`;
+    // That specific version should no longer exist
+    try {
+      await store.downloadBlob(
+        ctx,
+        ACCOUNT,
+        containerName,
+        name,
+        undefined,
+        originalLastModifiedIso
+      );
+      assert.fail("Should have thrown error for deleted specific version");
+    } catch (error) {
+      // Expected
+    }
+  });
+});
 
-      // Create version 1
-      const v1 = buildBlockBlob(ACCOUNT, containerName, name, "version1");
-      const created1 = await store.createBlob(ctx, v1);
+describe("LokiBlobMetadataStore - Versioning Enabled - deleteBlob comprehensive code path coverage @loki", () => {
+  let store: LokiBlobMetadataStore;
+  let disabledStore: LokiBlobMetadataStore;
+  let ctx: Context;
+  const containerName = "test-container";
 
-      // Create version 2
-      ctx.startTime = new Date(Date.now() + 100);
-      const v2 = buildBlockBlob(ACCOUNT, containerName, name, "version2");
-      await store.createBlob(ctx, v2);
+  beforeEach(async () => {
+    ctx = createContext();
+    // Versioning enabled
+    store = new LokiBlobMetadataStore("__test_db_blob__.json", false, true);
+    await store.init();
+    await store.createContainer(ctx, buildContainer(ACCOUNT, containerName));
 
-      // Delete version 1 specifically
+    // Versioning disabled
+    disabledStore = new LokiBlobMetadataStore(
+      "__test_db_blob_disabled__.json",
+      false,
+      false
+    );
+    await disabledStore.init();
+    await disabledStore.createContainer(
+      ctx,
+      buildContainer(ACCOUNT, containerName)
+    );
+  });
+
+  afterEach(async () => {
+    await store.close();
+    await store.clean();
+    await disabledStore.close();
+    await disabledStore.clean();
+  });
+
+  it("should throw error when versionId is provided with snapshot option @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    const created = await store.createBlob(ctx, blob);
+
+    // Create a snapshot
+    const snapshot = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+
+    try {
       await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-        versionId: created1.versionId
+        snapshot: snapshot.snapshot,
+        versionId: created.versionId
       });
+      assert.fail(
+        "Should have thrown error when versionId provided with snapshot"
+      );
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 400);
+      assert.ok(
+        error.message.includes(
+          "When deleting a blob version, you cannot specify a snapshot"
+        )
+      );
+    }
+  });
 
-      // Version 2 should still exist as current
-      const current = await store.downloadBlob(
+  it("should throw error when versionId is provided with deleteSnapshots option @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    const created = await store.createBlob(ctx, blob);
+
+    try {
+      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
+        deleteSnapshots: Models.DeleteSnapshotsOptionType.Include,
+        versionId: created.versionId
+      });
+      assert.fail(
+        "Should have thrown error when versionId provided with deleteSnapshots"
+      );
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 400);
+      assert.ok(
+        error.message.includes(
+          "When deleting a blob version, you cannot specify a snapshot"
+        )
+      );
+    }
+  });
+
+  it("should throw BlobNotFound when deleting non-existent blob @loki", async () => {
+    try {
+      await store.deleteBlob(
         ctx,
         ACCOUNT,
         containerName,
-        name
+        "non-existent-blob",
+        {}
       );
-      assert.strictEqual(
-        current.properties.contentLength,
-        Buffer.byteLength("version2")
+      assert.fail("Should have thrown BlobNotFound error");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+  });
+
+  it("should throw error when trying to use deleteSnapshots against a snapshot @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
+
+    // Create a snapshot
+    const snapshot = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+
+    try {
+      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
+        snapshot: snapshot.snapshot,
+        deleteSnapshots: Models.DeleteSnapshotsOptionType.Include
+      });
+      assert.fail(
+        "Should have thrown error when using deleteSnapshots against snapshot"
       );
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 400);
+      assert.ok(
+        error.message.includes("Invalid operation against a blob snapshot")
+      );
+    }
+  });
 
-      // Version 1 should be gone
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          "",
-          created1.versionId
-        );
-        assert.fail("Should have thrown error for deleted version");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
+  it("should delete specific version when versionId is provided @loki", async () => {
+    const name = `blob-${uuid()}`;
+
+    // Create version 1
+    const v1 = buildBlockBlob(ACCOUNT, containerName, name, "version1");
+    const created1 = await store.createBlob(ctx, v1);
+
+    // Create version 2
+    ctx.startTime = new Date(Date.now() + 100);
+    const v2 = buildBlockBlob(ACCOUNT, containerName, name, "version2");
+    await store.createBlob(ctx, v2);
+
+    // Delete version 1 specifically
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
+      versionId: created1.versionId
     });
 
-    it("should throw SnapshotsPresent when deleting base blob with snapshots (versioning enabled) @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
+    // Version 2 should still exist as current
+    const current = await store.downloadBlob(ctx, ACCOUNT, containerName, name);
+    assert.strictEqual(
+      current.properties.contentLength,
+      Buffer.byteLength("version2")
+    );
 
-      // Create snapshots
-      await store.createSnapshot(ctx, ACCOUNT, containerName, name);
-      await store.createSnapshot(ctx, ACCOUNT, containerName, name);
+    // Version 1 should be gone
+    try {
+      await store.downloadBlob(
+        ctx,
+        ACCOUNT,
+        containerName,
+        name,
+        "",
+        created1.versionId
+      );
+      assert.fail("Should have thrown error for deleted version");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+  });
 
-      try {
-        await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
-        assert.fail("Should have thrown SnapshotsPresent error");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 409);
-        assert.ok(error.message.includes("has snapshots"));
-      }
-    });
+  it("should throw SnapshotsPresent when deleting base blob with snapshots (versioning enabled) @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
 
-    it("should throw SnapshotsPresent when deleting base blob with snapshots (versioning disabled) @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await disabledStore.createBlob(ctx, blob);
+    // Create snapshots
+    await store.createSnapshot(ctx, ACCOUNT, containerName, name);
+    await store.createSnapshot(ctx, ACCOUNT, containerName, name);
 
-      // Create snapshots
-      await disabledStore.createSnapshot(ctx, ACCOUNT, containerName, name);
-      await disabledStore.createSnapshot(ctx, ACCOUNT, containerName, name);
-
-      try {
-        await disabledStore.deleteBlob(ctx, ACCOUNT, containerName, name, {});
-        assert.fail("Should have thrown SnapshotsPresent error");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 409);
-        assert.ok(error.message.includes("has snapshots"));
-      }
-    });
-
-    it("should mark blob as non-current when deleting base blob without snapshots (versioning enabled) @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      const created = await store.createBlob(ctx, blob);
-
-      // Delete the blob (no snapshots exist)
+    try {
       await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
+      assert.fail("Should have thrown SnapshotsPresent error");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 409);
+      assert.ok(error.message.includes("has snapshots"));
+    }
+  });
 
-      // Blob should be marked as non-current, not physically deleted
-      try {
-        await store.downloadBlob(ctx, ACCOUNT, containerName, name);
-        assert.fail("Should have thrown error for non-current blob");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
+  it("should throw SnapshotsPresent when deleting base blob with snapshots (versioning disabled) @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await disabledStore.createBlob(ctx, blob);
 
-      // But should still be accessible by version ID
-      const versionedBlob = await store.downloadBlob(
+    // Create snapshots
+    await disabledStore.createSnapshot(ctx, ACCOUNT, containerName, name);
+    await disabledStore.createSnapshot(ctx, ACCOUNT, containerName, name);
+
+    try {
+      await disabledStore.deleteBlob(ctx, ACCOUNT, containerName, name, {});
+      assert.fail("Should have thrown SnapshotsPresent error");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 409);
+      assert.ok(error.message.includes("has snapshots"));
+    }
+  });
+
+  it("should mark blob as non-current when deleting base blob without snapshots (versioning enabled) @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    const created = await store.createBlob(ctx, blob);
+
+    // Delete the blob (no snapshots exist)
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {});
+
+    // Blob should be marked as non-current, not physically deleted
+    try {
+      await store.downloadBlob(ctx, ACCOUNT, containerName, name);
+      assert.fail("Should have thrown error for non-current blob");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+
+    // But should still be accessible by version ID
+    const versionedBlob = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      "",
+      created.versionId
+    );
+    assert.strictEqual(
+      versionedBlob.properties.contentLength,
+      Buffer.byteLength("content")
+    );
+  });
+
+  it("should physically delete blob when deleting base blob without snapshots (versioning disabled) @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    const created = await disabledStore.createBlob(ctx, blob);
+
+    // Delete the blob (no snapshots exist)
+    await disabledStore.deleteBlob(ctx, ACCOUNT, containerName, name, {});
+
+    // Blob should be completely gone
+    try {
+      await disabledStore.downloadBlob(ctx, ACCOUNT, containerName, name);
+      assert.fail("Should have thrown error for deleted blob");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+
+    // Should not be accessible by version ID either (versioning disabled)
+    try {
+      await disabledStore.downloadBlob(
         ctx,
         ACCOUNT,
         containerName,
@@ -3885,305 +3157,272 @@ describe("LokiBlobMetadataStoreVersioning", () => {
         "",
         created.versionId
       );
-      assert.strictEqual(
-        versionedBlob.properties.contentLength,
-        Buffer.byteLength("content")
-      );
+      assert.fail("Should have thrown error for deleted blob by version");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+  });
+
+  it("should delete individual snapshot when targeting specific snapshot @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
+
+    // Create multiple snapshots
+    const snapshot1 = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    ctx.startTime = new Date(Date.now() + 100);
+    const snapshot2 = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+
+    // Delete first snapshot specifically
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
+      snapshot: snapshot1.snapshot
     });
 
-    it("should physically delete blob when deleting base blob without snapshots (versioning disabled) @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      const created = await disabledStore.createBlob(ctx, blob);
+    // Base blob should still exist
+    const baseBlob = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    assert.strictEqual(
+      baseBlob.properties.contentLength,
+      Buffer.byteLength("content")
+    );
 
-      // Delete the blob (no snapshots exist)
-      await disabledStore.deleteBlob(ctx, ACCOUNT, containerName, name, {});
+    // Second snapshot should still exist
+    const snap2 = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      snapshot2.snapshot
+    );
+    assert.strictEqual(
+      snap2.properties.contentLength,
+      Buffer.byteLength("content")
+    );
 
-      // Blob should be completely gone
-      try {
-        await disabledStore.downloadBlob(ctx, ACCOUNT, containerName, name);
-        assert.fail("Should have thrown error for deleted blob");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
+    // First snapshot should be gone
+    try {
+      await store.downloadBlob(
+        ctx,
+        ACCOUNT,
+        containerName,
+        name,
+        snapshot1.snapshot
+      );
+      assert.fail("Should have thrown error for deleted snapshot");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+  });
 
-      // Should not be accessible by version ID either (versioning disabled)
-      try {
-        await disabledStore.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          "",
-          created.versionId
-        );
-        assert.fail("Should have thrown error for deleted blob by version");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
+  it("should delete base blob and all snapshots when deleteSnapshots=include (versioning disabled) @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await disabledStore.createBlob(ctx, blob);
+
+    // Create snapshots
+    const snapshot1 = await disabledStore.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    const snapshot2 = await disabledStore.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+
+    // Delete blob and all snapshots
+    await disabledStore.deleteBlob(ctx, ACCOUNT, containerName, name, {
+      deleteSnapshots: Models.DeleteSnapshotsOptionType.Include
     });
 
-    it("should delete individual snapshot when targeting specific snapshot @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
+    // Base blob should be gone
+    try {
+      await disabledStore.downloadBlob(ctx, ACCOUNT, containerName, name);
+      assert.fail("Should have thrown error for deleted blob");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
 
-      // Create multiple snapshots
-      const snapshot1 = await store.createSnapshot(
+    // All snapshots should be gone
+    try {
+      await disabledStore.downloadBlob(
         ctx,
         ACCOUNT,
         containerName,
-        name
+        name,
+        snapshot1.snapshot
       );
-      ctx.startTime = new Date(Date.now() + 100);
-      const snapshot2 = await store.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
+      assert.fail("Should have thrown error for deleted snapshot1");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
 
-      // Delete first snapshot specifically
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-        snapshot: snapshot1.snapshot
-      });
-
-      // Base blob should still exist
-      const baseBlob = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-      assert.strictEqual(
-        baseBlob.properties.contentLength,
-        Buffer.byteLength("content")
-      );
-
-      // Second snapshot should still exist
-      const snap2 = await store.downloadBlob(
+    try {
+      await disabledStore.downloadBlob(
         ctx,
         ACCOUNT,
         containerName,
         name,
         snapshot2.snapshot
       );
-      assert.strictEqual(
-        snap2.properties.contentLength,
-        Buffer.byteLength("content")
-      );
+      assert.fail("Should have thrown error for deleted snapshot2");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+  });
 
-      // First snapshot should be gone
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          snapshot1.snapshot
-        );
-        assert.fail("Should have thrown error for deleted snapshot");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
+  it("should delete snapshots only and mark base blob as non-current when deleteSnapshots=include (versioning enabled) @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    const created = await store.createBlob(ctx, blob);
+
+    // Create snapshots
+    const snapshot1 = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    const snapshot2 = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+
+    // Delete blob and all snapshots
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
+      deleteSnapshots: Models.DeleteSnapshotsOptionType.Include
     });
 
-    it("should delete base blob and all snapshots when deleteSnapshots=include (versioning disabled) @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await disabledStore.createBlob(ctx, blob);
+    // Base blob should be marked as non-current
+    try {
+      await store.downloadBlob(ctx, ACCOUNT, containerName, name);
+      assert.fail("Should have thrown error for non-current blob");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
 
-      // Create snapshots
-      const snapshot1 = await disabledStore.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-      const snapshot2 = await disabledStore.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-
-      // Delete blob and all snapshots
-      await disabledStore.deleteBlob(ctx, ACCOUNT, containerName, name, {
-        deleteSnapshots: Models.DeleteSnapshotsOptionType.Include
-      });
-
-      // Base blob should be gone
-      try {
-        await disabledStore.downloadBlob(ctx, ACCOUNT, containerName, name);
-        assert.fail("Should have thrown error for deleted blob");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
-
-      // All snapshots should be gone
-      try {
-        await disabledStore.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          snapshot1.snapshot
-        );
-        assert.fail("Should have thrown error for deleted snapshot1");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
-
-      try {
-        await disabledStore.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          snapshot2.snapshot
-        );
-        assert.fail("Should have thrown error for deleted snapshot2");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
-    });
-
-    it("should delete snapshots only and mark base blob as non-current when deleteSnapshots=include (versioning enabled) @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      const created = await store.createBlob(ctx, blob);
-
-      // Create snapshots
-      const snapshot1 = await store.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-      const snapshot2 = await store.createSnapshot(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-
-      // Delete blob and all snapshots
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-        deleteSnapshots: Models.DeleteSnapshotsOptionType.Include
-      });
-
-      // Base blob should be marked as non-current
-      try {
-        await store.downloadBlob(ctx, ACCOUNT, containerName, name);
-        assert.fail("Should have thrown error for non-current blob");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
-
-      // All snapshots should be gone
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          snapshot1.snapshot
-        );
-        assert.fail("Should have thrown error for deleted snapshot1");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
-
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          snapshot2.snapshot
-        );
-        assert.fail("Should have thrown error for deleted snapshot2");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
-
-      // But base blob should still be accessible by version ID
-      const versionedBlob = await store.downloadBlob(
+    // All snapshots should be gone
+    try {
+      await store.downloadBlob(
         ctx,
         ACCOUNT,
         containerName,
         name,
-        "",
-        created.versionId
+        snapshot1.snapshot
       );
-      assert.strictEqual(
-        versionedBlob.properties.contentLength,
-        Buffer.byteLength("content")
+      assert.fail("Should have thrown error for deleted snapshot1");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+
+    try {
+      await store.downloadBlob(
+        ctx,
+        ACCOUNT,
+        containerName,
+        name,
+        snapshot2.snapshot
       );
+      assert.fail("Should have thrown error for deleted snapshot2");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+
+    // But base blob should still be accessible by version ID
+    const versionedBlob = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name,
+      "",
+      created.versionId
+    );
+    assert.strictEqual(
+      versionedBlob.properties.contentLength,
+      Buffer.byteLength("content")
+    );
+  });
+
+  it("should delete only snapshots when deleteSnapshots=only @loki", async () => {
+    const name = `blob-${uuid()}`;
+    const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
+    await store.createBlob(ctx, blob);
+
+    // Create snapshots
+    const snapshot1 = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    const snapshot2 = await store.createSnapshot(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+
+    // Delete only snapshots
+    await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
+      deleteSnapshots: Models.DeleteSnapshotsOptionType.Only
     });
 
-    it("should delete only snapshots when deleteSnapshots=only @loki", async () => {
-      const name = `blob-${uuid()}`;
-      const blob = buildBlockBlob(ACCOUNT, containerName, name, "content");
-      await store.createBlob(ctx, blob);
+    // Base blob should still exist and be current
+    const baseBlob = await store.downloadBlob(
+      ctx,
+      ACCOUNT,
+      containerName,
+      name
+    );
+    assert.strictEqual(
+      baseBlob.properties.contentLength,
+      Buffer.byteLength("content")
+    );
 
-      // Create snapshots
-      const snapshot1 = await store.createSnapshot(
+    // All snapshots should be gone
+    try {
+      await store.downloadBlob(
         ctx,
         ACCOUNT,
         containerName,
-        name
+        name,
+        snapshot1.snapshot
       );
-      const snapshot2 = await store.createSnapshot(
+      assert.fail("Should have thrown error for deleted snapshot1");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
+
+    try {
+      await store.downloadBlob(
         ctx,
         ACCOUNT,
         containerName,
-        name
+        name,
+        snapshot2.snapshot
       );
-
-      // Delete only snapshots
-      await store.deleteBlob(ctx, ACCOUNT, containerName, name, {
-        deleteSnapshots: Models.DeleteSnapshotsOptionType.Only
-      });
-
-      // Base blob should still exist and be current
-      const baseBlob = await store.downloadBlob(
-        ctx,
-        ACCOUNT,
-        containerName,
-        name
-      );
-      assert.strictEqual(
-        baseBlob.properties.contentLength,
-        Buffer.byteLength("content")
-      );
-
-      // All snapshots should be gone
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          snapshot1.snapshot
-        );
-        assert.fail("Should have thrown error for deleted snapshot1");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
-
-      try {
-        await store.downloadBlob(
-          ctx,
-          ACCOUNT,
-          containerName,
-          name,
-          snapshot2.snapshot
-        );
-        assert.fail("Should have thrown error for deleted snapshot2");
-      } catch (error) {
-        assert.strictEqual(error.statusCode, 404);
-      }
-    });
+      assert.fail("Should have thrown error for deleted snapshot2");
+    } catch (error) {
+      assert.strictEqual(error.statusCode, 404);
+    }
   });
 });
